[roll] Roll fuchsia [overnet] Fix router bug
The overnet router was keeping track of which nodes a route passed through,
but it was inserting the wrong node-id into the list.  As a result, if node
A had links to both B and C, it would think it could get to B both directly
_and_ via C, and similarly it would think it could get to C via B.  While
these loops should not cause actual problems routing (since the shortest path
is always used, and the indirect path is always slower than the direct path),
it is conceivable that in certain exception cases (as as links are shutting
down), some pathological behavior could be revealed. At the least, it's clearly
an error.

(Fixing this bug revealed an additional bug in one of the tests, one that had
been incorrectly passing.)

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717757
Original-Revision: 908cf576215b7c7a20356f528dadd75e8c9c0f1e
GitOrigin-RevId: 7810b1dcc025d09105ffe39524674026c67f61e8
Change-Id: I32f2259bcf70281b6d35530867b9ca83df41c9a3