[roll] Roll fuchsia [magma] Disable constrained mapping tests with HWASAN enabled
PlatformBuffer.MapConstrained and PlatformBuffer.Padding rely on mapping
a VMO into the first 4GB of address space, which may not be possible if
the HWASAN shadow stack is allocated there. So we need to disable the
tests when HWASAN is enabled.

Original-Bug: 108382
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719090
Original-Revision: 21ad6e84f9ced2655eea688537065ba87b435687
GitOrigin-RevId: 9f1773ff6183fc1f805153c89d10dce86466096a
Change-Id: Id2cfcb719ab65a5c6c26eae11aae201b74fa0ac3