[roll] Roll fuchsia [build] Add clang-doc templates and wrapper.
Creates some GN templates for invoking clang-doc.

Adds a script to synthesize a header file so the headers that should be
documented can be referenced by the build.

Adds a script to wrap a clang-doc invocation, redirecting the output to
a temporary directory that can be zipped (the build prohibits actions
writing output files that aren't known in advance).

I tested these manually but they are not currently hooked up. They will
be used in a followup by the tool that consumes the clang-doc output.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711903
Original-Revision: 03515eacb67d4cf2c442118a9a6a248517c2a8aa
GitOrigin-RevId: e5618648b6f737fe3f570098e1e200e55cd6a57c
Change-Id: Iea0e4c326b60db7817bb4231d608639b13b4c4c2