[roll] Roll fuchsia [storage][conformance] Enforce max read size of MAX_TRANSFER_SIZE.
Enforces the documented limit in fuchsia.io of MAX_TRANSFER_SIZE for any
transfer buffer. This is enforced on write by the fidl client code, but
for read the server should reject requests for larger sizes.

This found a bug in the cppvfs server where it would try to send an
oversized reply and hang the fidl connection.

Original-Bug: 82674
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718999
Original-Revision: 650fc45d668704b2acdfcb024ef596ec9b231c06
GitOrigin-RevId: f99540592b13c0ab6e1af98a21a58a1e1b94ed31
Change-Id: I211a16bbbe848369732515d039494051a208eaa5