[roll] Roll fuchsia [hypervisor][x86] Fix invvpid after vmxoff in DirectGuest
The address space associated with a direct mode guest may be destroyed
after the final guest has been destroyed, and therefore VMX has been
turned off. This will then cause an exception to be raised.

To avoid this, we reset the VPID associated with the address space on
destruction. This is safe as on subsequent creation of a guest, we will
always issue an invept, and then on create of a VCPU we will issue an
invvpid.

Original-Bug: 95763
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719215
Original-Revision: 7e41ef9ec8651b449dde2fdab93e0b0fcd2c4fa0
GitOrigin-RevId: 2bfd4a310be40e4ecc26d338505551861371e8b1
Change-Id: Iafe5fa41b628be52424aba223deab33af427c94a