[roll] Roll fuchsia [fit] Fix null_target::ops duplicate symbol errors
A recent change removed the template parameters from the fit::function
null target class. If null_target is not a template, null_target::ops
must be defined in a source file in C++14. In C++17, it could be defined
as an inline variable either in the class or elsewhere in the header.

To avoid introducing a new .cc file for null_target::ops, make
null_target a template again, but only instantiate it once. This
prevents potential multiple definition errors for null_target::ops.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719306
Original-Revision: f73d8db89708db77a465d7757777cf1c64899d3f
GitOrigin-RevId: 31b08b72e846d5da3ab8b37b5453a6cc40221f7b
Change-Id: Ia282424e1ea633903d7e1240d5784c5dc6334749