[roll] Roll fuchsia [f2fs] Count the i_block in the same way as a generic filesystems.
For compatibility, change the in-memory i_block to 512-byte sectors, the same as a generic filesystems.
In addition, the inode block is excluded from the in-memory i_block.
On the other hand, we calculate the block count in the on-disk i_block and keep the inode block included.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708082
Original-Revision: 1e7f54a82031430a86a3cddf8dcf82b7f0346292
GitOrigin-RevId: f0c7a47ce083f0076b64fcc8f3baf2cfa64fe828
Change-Id: I2cd812d5886d386de6fe250f4618e2527b1c4e3b