[roll] Roll fuchsia [fuchsia-map] Create component and ffx plugin
The component called process_explorer gets all the processes on a running Fuchsia device, and for each of them it retrieves information about all the handles in the process. The data is then packaged as a string in JSON format and sent over a socket.

Component tested by using the Realm Builder library and mocking fuchsia.kernel.RootJob.

The ffx plugin, which is also called process_explorer, starts the component, receives the data about all processes as a string and prints it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699522
Original-Revision: e770ad790dba49191d3a114fe7fa35bdc1e02546
GitOrigin-RevId: ec2fac754dda20e6266bba172708dead2f4ef71e
Change-Id: I13d739e6f5d86e5346e236507fc7a0ed0eda5630