[roll] Roll fuchsia [bt][fastpair] Support pairing over LE & BR/EDR
Context: The remote device will initiate Fast Pair over LE. When the
peer is first discovered by the system, it is assigned a unique PeerId.
The remote device will then try to pair with us. In some cases, this is
over BR/EDR and as such, the system will assign a new unique PeerId to
the same remote device. Because the ids differ for the same peer,
`bt-fastpair-provider` will reject the pairing request.

Update the PairingManager to try and match the PeerId to an active Fast
Pair pairing procedure. This is a best-effort operation and likely to
succeed. It is unlikely that there are multiple ongoing pairing
procedures.

File fxbug.dev/107780 to see if we can simplify the implementation such
that only one Fast Pair pairing procedure is supported at a time.

Original-Bug: 107780
Test: fx test bt-fastpair-provider-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716230
Original-Revision: c15459294833323304ff73ad12be6036f1d1dee5
GitOrigin-RevId: 30d871d8be94480585d1be01b6320b91b6e30ca8
Change-Id: I3b15bf421b99aa63a086b8041396b9811c4d1f85