[roll] Roll fuchsia [fidl][dynsuite] unknown interactions server tests.
Implement unknown interactions tests for the server in the DynSuite.
These tests are enabled for LLCPP, Natural Types, and Rust. For Go, they
are disabled because we have not implemented unknown interactions
support. For HLCPP they are disabled because unknown interactions
support is pending but not yet committed.

Original-Bug: 88366

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709902
Original-Revision: ae9dc16e816393afdf2453df96aa24e47256174f
GitOrigin-RevId: 29aac30263f4ffb7d76d41969f2da21faac55dd0
Change-Id: I79264d2086f416dd30eecdaab76873a49f1f3744