[roll] Roll fuchsia [thermd] Ensure that logs are always printed on termination
This is workaround for fxb/97657 where logs aren't printed if the
component exits before archivist starts. The workaround is to wait
around whenever we hit an error, which in the case of thermd is any
termination path.

Note that this makes marking thermd as a critical component less useful,
since it will only terminate on a crash, and we may want to do that in
the future. For now, it's more important to log *why* failures occur
than to reboot the device on failure.

Original-Bug: 97657
Original-Bug: 108619
Original-Bug: 108624

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719803
Original-Revision: 86f3c8ee613d8a02c1e919d2c93fc62c8969c81d
GitOrigin-RevId: 19876eafaf4cc81f40a2aff8d81dae3592351f8b
Change-Id: I4c39ed59d043f1dc32ad05c72c94308a00129da6