[roll] Roll fuchsia [netstack3] Track Address State in Bindings
Change [1/2] in supporting the `WatchAddressAssignmentState` FIDL API.
All relevant integration tests are enabled in the follow up CL.

Propagate updates in `AddressAssignmentState` to the
`AddressStateProvider` Worker when handling `IpDeviceEvents` published
by Core. Updates are sent via an `Unbounded` mpsc channel, which exposes
a synchronized non-blocking send API (required since Core is no-async).

As a result of this change, the initialization logic of the
`AddressStateProvider` worker changed slightly. AddressInfo is added
to `ctx` before the address is added to Core, which ensures the sender
is inplace when Core generates the "Address Added" `IpDeviceEvent`. The
`AddressStateProvider` worker now adds the address to Core as part of
startup.

Original-Bug: 105011

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715682
Original-Revision: 776d9d146fc8d93536a0eff85b2cf4c43b726e0e
GitOrigin-RevId: 59f886030a67443195953965501024c77bcb6f80
Change-Id: I8b71a23333438a0c1455b1b1f585c6ad82051ece