[roll] Roll fuchsia [dash-launcher] Add support for custom namespace layouts
This CL adds a new flag to `ffx component explore` called
`-l` or `--layout` that allows changing the namespace
layout given to the dash shell.

The current options are `nested` and `namespace`.

By default, the namespace layout is nested. This means that
instance directories are nested in subdirs at the root
(e.g - component namespace is under /ns, etc.)

When set to `namespace`, the instance namespace is the root.
This most closely resembles what a component instance would
see if it queried its own namespace. This layout is more
compatible with running tools.

To achieve this, we added a new parameter to the Launcher
protocol for selecting the namespace layout.

Original-Bug: 106028
Test: fx test dash_launcher
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719047
Original-Revision: a8027868f8897255fd7f899c1236763ecd7e54ba
GitOrigin-RevId: 246c604ba7aaf1d3be07c78dc5f2ef91e8e34527
Change-Id: I965305b539b7f53e4449be04533368698af1cf9a