[roll] Roll fuchsia [net-types] Build Cargo features in GN
Create a GN target for the no-std version of net-types (with the "std"
feature disabled). Ensure that both the target with this feature enabled
and the target with it disabled are built and tested in GN. While the
"std" feature is used by all in-tree code, future in-tree code may
disable it, and once we publish on crates.io, external dependents may
disable the it. Thus, it's important to prevent it from bitrotting.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720041
Original-Revision: e6f5086a943c8023f007c2e5906c871f9bfce75a
GitOrigin-RevId: dfedde62e65613d3b78acbdaf69f9c986b383c27
Change-Id: I5f56f5bba1ca16addab8aa1faa81e3dd404afaff