[roll] Roll fuchsia [policy] Cleanup the structure.
The //src/security/policy directory is our highest traffic directory.
We need to make it as easy as possible to add a new policy without
being confronted by a thousand small files. Almost always the user
is looking for the cfv2 allowlist and very rarely they are looking
for something to do with appmgr. Pkgfs is also a very different
type of allowlist than the other ones and so it deserves its own
distinct directory. This cleanup should make it easier for a new
Fuchsia Engineer who has never worked with policies before just jump
in make a few line edit and be ready to go.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720457
Original-Revision: 57f99290284f8a96889ff4098752e08d77e96724
GitOrigin-RevId: 5b1bfc7748212b316b5b7ec656ac3daa10f9e13f
Change-Id: I9a3668de2d937faf4fe925a655608f3173c18c9b