[roll] Roll fuchsia [zerocopy] Build most Cargo features in GN
Create one GN target for each Cargo feature (with the exception of
"simd-nightly", which requires unstable compiler features which are not
enabled in our build). Ensure that all targets - even those which are
not used elsewhere in our build - are built and tested. This helps to
prevent Cargo features which are not used in-tree but which are
available on crates.io do not bitrot.

Also simplify the way that we test feature-flagged code. Previously, all
features were tested when `cfg(test)` was enabled. In this commit, we
change that behavior so that feature-flagged code is only tested when
that feature is enabled. This somewhat simplifies our `cfg`s.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720403
Original-Revision: a3f7d02fa0ca01263a1a5dd72faed286109e8701
GitOrigin-RevId: daf6c4e932bbf2e03e6e4724742a25141ef6c8fe
Change-Id: I33427a3cd5531474febfec30c6f3d0e801320e80