[roll] Roll fuchsia [x86][idle] Don't call Scheduler::Reschedule
Don't call Scheduler::Reschedule when dropping the preempt disable
after waiting up in the idle loop.  Calling Reschedule is a bit of a
waste of effort.  It is going to require us to bounce through the
thread lock, and we know it is not actually going to do anything as we
have explicitly disabled local preemption.

The only thing it is actually going to do is set the local CPU
preemption flag.  We can do that ourselves much more quickly with a
simple relaxed atomic or.  There is no need to make things more
complicated than that.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717179
Original-Revision: 2f774a26ed359d200f9506db920a151a83902bd2
GitOrigin-RevId: b2cd875e8c28dfe23c63a18a45b4ed33f9eff14a
Change-Id: I039a7a71362316ef56280aefb6d3cbf1e5de6006