[roll] Roll fuchsia [netstack] Return error on addr/ep IP mismatch
On Linux, it's not possible to connect to an address with a different IP
version than the endpoint's version. Match this behavior for both V4 and
V6 (the previous code only did so for V4). In addition, add test
coverage for the code path in which we try to connect to a mismatched
IP address when the socket already has an error. This coverage is in
preparation for a later commit that refactors the handling of mismatched
IP addr errors, which (absent coverage) can introduce panics in the
Netstack.

Original-Fixed: 108755
Original-Bug: 104640
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719646
Original-Revision: d9db035f5b33b5cd2a1060c9fbab9b660329e21f
GitOrigin-RevId: ce9e296a86c1923244cd694eb90e4e5e4a419e87
Change-Id: Ifd56cfdc04eb606fff3e1b8fd1ea133f8a544f40