[roll] Roll fuchsia [ctf] //sdk/cts -> //sdk/ctf
Note that this retains build rules relied on by infra, as well as
adding new build rules for those users to migrate to.  As a next step,
out of tree users should migrate to the new build rules.

Steps:

1) sed --in-place -e "s|sdk/cts|sdk/ctf|g" $(csearch --nostats -l \
     --corpus=fuchsia "sdk/cts")
2) cd fuchsia/sdk
3) git mv cts ctf

4) Re-add sdk/cts build rules depended on from OOT and hand-tweak
   to point to sdk/ctf.

5) Change tools/fidl templates to point to ctf/ instead of cts/:
   sed --in-place -e "s|cts/tests|ctf/tests|g" \
      $(csearch --nostats -l --corpus=fuchsia \
        "cts/tests/pkg/fidl/cpp/test/handle_util.h")

Note that this means that a bunch of output now has duplicated
variants in cts and ctf directories.  We should migrate OOT
dependencies off of cts variants and delete them as soon as is
feasible.

Of specific note is that the following files are now being
generated twice:

- cts/test_manifest.json is also ctf/test_manifest.json
- cts/host_test_manifest.json is also ctf/host_test_manifest.json
- tests/cts_unstripped_binaries_list is also
   tests/ctf_unstripped_binaries_list

Original-Bug: 93414

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719670
Original-Revision: bc08acfdb1f883fa100257d8f267cc2b2a594794
GitOrigin-RevId: 512c4d4750375d923c1e7eef8622fbe294a15af4
Change-Id: I3c817d61c1181f17d75bc021955901aa460830c4