[roll] Roll fuchsia [ffx][fuzz] Standardize filesystem layout
This CL adjusts the layout of the files written by the `ffx fuzz`
plugin. First, it requires an `-o` or `--output` flag to be provided to
the `attach` subcommand or a `fuzz.output` value to be set by
`ffx config`. This value must be a path to a writable directory.

Depending on the subcommands executed, the plugin may write to the
following subdirectories under the `<output>` directory:

 * `<output>/logs/`: stdout, stderr, and syslog from the fuzzer.
 * `<output>/artifacts/`: test inputs returned by the fuzzer.
 * `<output>/seed-corpus/`: set of fixed inputs used to initialize the
   fuzzer.
 * `<output>/corpus/`: set of inputs generated by the fuzzer that
   increased code coverage.

The plugin will no longer default to dumping files in the current
working directory.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716447
Original-Revision: 4f80ce739fc29a4c1195aa7dfdedd52cff27ec4f
GitOrigin-RevId: 473ba30cdd94c0e57fea6885db0de81fe0111ff1
Change-Id: Icc5d3236e2fd23feb0f25dd3f838ce26ec1f2292