[roll] Roll fuchsia [virtualization] Bind to balloon via guest protocol
This change moves connecting to the balloon controller into the guest
protocol. The goal is to have the VMM only expose a single client API,
and a single control plane API.

This also simplifies VMM lifecycle. Before this change we had to ensure
that the VMM was running before binding to the BalloonController (since
it doesn't make sense without a device), but now that's implicitly
enforced by using the guest protocol served by the VMM.

Original-Bug: 104989
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720504
Original-Revision: a69037865b29cdf6879937bff5e615f433036db8
GitOrigin-RevId: 3493489d3b87cf9f6ab33dd826e8126caca28b4a
Change-Id: I8d7d299d45e58797ce9caa7042f4a200207cd0d9