[roll] Roll fuchsia [eager_package_config] Remove test config.
If a build includes the unit test, the test eager package config will
also be included in the build, affecting the real omaha-client and
pkg-resolver behavior.

The purpose of this test config is to test invocation of the script from
GN, now that we have a real eager package config for ota test, this
config can be removed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/721463
Original-Revision: 7297c685234599348e3a26ca633d4eeb46788bec
GitOrigin-RevId: 89266c038249a43674d63394142ff7e688f2087b
Change-Id: Icce9acd89ed7d3e693c5189905c52d694e037a41