[roll] Roll fuchsia [docgen] Add an end-to-end test.
Adds some headers that are run through the documentation infrastructure
and compared to golden versions of the generated documentation.

Check in golden files representing the current state of the
documentation generator. Note that these generate .md for devsite to
serve which is not always handled by the Gerrit previewer.

Currently this is not run by default. Once the GN patch to improve
compilation database generation is landed we can enhance this.

Add the capability of the build templates to emit a directory of files.
This can be used because the golden file infrastructure requires that we
know the output names of all files anyway.

Improve error messages of clang doc failures.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/721179
Original-Revision: 576229d91716f587d882f82e48c9d6f9d3593adf
GitOrigin-RevId: c034643a13668e59d9d800c70397f1c75991313e
Change-Id: Ifc84f2d57bc32f361304e31af4af98e3d3806066