[roll] Roll fuchsia [emu] Add --config to ffx emu show
This CL adds the "--config" flag to the `ffx emu show` subcommand. This
flag outputs the FlagData/Launch Config in the same format as that taken
as input by the "--config" flag for `ffx emu start`. The idea here is
that a user can start an emulator, copy the config using this new flag,
and then feed that config back to a new emulator instance in the future
instead of having to re-build the configuration. This is especially
useful when custom flags are added using the "--edit" functionality.

This CL also moves the existing default to the "--raw" flag, and
repurposes the "--all" flag to print all of the formatted data segments
(excluding the raw output). The raw output remains the default.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/721722
Original-Revision: 9f4c0acb2cae58c297b00767b46012b8b0600494
GitOrigin-RevId: d992df705745ce9334d83c1238728ee4dd8fd176
Change-Id: Id544553499466d16fff7981b81a3c56a9e616c10