[roll] Roll fuchsia [netstack/bench] Don't double record loopback
Currently the netstack iperf3 benchmarks mark the benchmark over
loopback as both a "send" and "recv" benchmark. This is somewhat
misleading because all other metrics use "send" to indicate a test where
the device is sending to the host, and "recv" for the reverse, whereas
in the loopback case "send" and "recv" are from the same test. This also
means that the same metric gets recorded twice.

Instead, record results from iperf3 over loopback just once, and label
each benchmarking setup as one of "send" (DUT-to-host), "recv"
(host-to-DUT), or "loopback" (on-device).

Original-Bug: 108749
Test: fx test --e2e host_x64/netstack_iperf_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720035
Original-Revision: 59895c6c399b654dfe2c8f55d792148142fc58c9
GitOrigin-RevId: cc970e426563e07312a103d01ab1044cb1b81a34
Change-Id: I547cf27a505505304882da813bd6da515a2a18a7