[roll] Roll fuchsia [e2e] Migrate to CF v2 in package_manager_test
This is a reland of https://fxrev.dev/702023. Different to the
original change we switch to using `run-test-suite` in the target
device shell instead of `run` which v1 used. Additionally we
trivially convert the example programs to tests so we can use
`run-test-suite`.

Original-Bug: fxb/103405

Test: fx test --e2e package_manager_test
Multiply: package_manager_test: 100

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/721985
Original-Revision: e5592919825d707574c17321127362a9c9313fe4
GitOrigin-RevId: 4263b3bc31553c9868250b8194a98c4c2283ab97
Change-Id: I7d7abeca6a2a2d34b0422613e5c75fa6ce8cf76f