[roll] Roll fuchsia [kernel][vm] Restrict AddNewPagesLocked
Forbids the CanOverwriteContent::NonZero option from being passed
to AddNewPagesLocked, removing the need for it to populate a list of
released pages. None of the existing users of AddNewPagesLocked ever
passed in a non-null list, and this path was never exercised.

This simplifies future changes where content other than vm_page_t might
be pre-existing, which would not be able to be placed in a linked list.

Original-Bug: 60238
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717586
Original-Revision: fc7868fae8e5be24686695fd6c0188e7ef8262c2
GitOrigin-RevId: f3586007a760694d23734352bd376b488cc07954
Change-Id: Ie0df8b93cae5e63e1dab169fc51693a4b7d30f5f