[roll] Roll fuchsia [netstack3] Implement Clone/Clone2 for TCP sockets
This diverges from the UDP implementation in that it runs all
operations for a TCP socket on the same executor. When a new socket is
created via Clone/Clone2, the requests for that socket are merged in
with the requests from the existing socket. This has the advantage of
not requiring external synchronization (since no other threads are
touching the same socket).

Original-Bug: 108728
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720889
Original-Revision: 7edb7cc96a4fa16bedf0ee82d8c8bb82e2d764fa
GitOrigin-RevId: 10aff606043837bf165bdae2fbd34f304bc4e04a
Change-Id: I9851d1567e359669538445c48d8343785d2021fa