[roll] Roll fuchsia [devicetree] Path utilities
In preparation for devicetree::Scan, we introduce some utilities
for matchers to use, which allow them to compare paths, and resolve
paths that contain aliases.

PathResolver will be constructed by the scan method, with the
set of aliases available at the time. The ResolvedPath is able
to communicate to the caller whether the path can be resolved
at this time(e.g. we haven reached aliases node yet, and we
try to resolve an alias) and when a path can be resolved,
whether it was successfully resolved or not(the alias is
not found, the alias is empty, etc).

The ComparePath is similar to std::equals for two ranges,
except that it communicates it results as a pair of iterators
to the comparing ranges. This can be used to compare two disjoint
ranges that represent a full range, without additional lookups.
E.g.
   aliases/rest_of_path
   -> /some_absolute_path , rest_of_path

A caller may want to know if the alias absolute path matches
and if the rest of the path matches as well.

Test: fx test -v -o device-tree-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717523
Original-Revision: 839aca62872810c00a258d0762885056d591ff95
GitOrigin-RevId: 3fc43090d1c785f8961213a32938c145cb6bc5a3
Change-Id: I7e432682deb6b157e9e4464c3c3d7526506044a8