[roll] Roll fuchsia [fidl][hlcpp] unknown interactions server.
Update HLCPP server to be able to handle unknown interactions. The API
used on the server is shared with the client, which means this API is
slightly suboptimal for the server.

For flexible methods, the response callback accepted on the server takes
the Result union, regardless of whether the method uses error syntax. If
the server tries to respond with the transport_err variant, the response
callback panics.

For handling unknown method calls, a handle_unknown_method method is
provided. To prevent this from being called on the Proxy types, this
method is declared as protected, and the implementation provided on the
proxy always panics.

Original-Bug: 88366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707223
Original-Revision: 440c26e0268cb292fe199f2c62452305aa389084
GitOrigin-RevId: 28da0e8d1f41ee533f5b71bf3b0dca9d560901b8
Change-Id: I12d71bf4a6d10d8d8efea7528be543f35d30686b