[roll] Roll infra/recipes [test_owners] Ignore invalid includes
The test-owners builder is intended to be a best effort at collecting
ownership data. It's not intended to be the place where we validate
OWNERS files, so it should be resilient to invalid OWNERS files, in
particular invalid `include` statements, which currently cause the
builder to become broken.

Original-Bug: 108186
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/722582
Original-Revision: 9490c279e4744796c1d2cd6ed6377a311563dbba
GitOrigin-RevId: d61e303a5579f5b1a36f5dc59053003a5e69019e
Change-Id: Ie167268eaf6ad615ca2ed70e2859df48267b47a3