[roll] Roll fuchsia [blackout] add response to test method
Add an expected response to the test fidl method on the controller.
This will allow the harness to confirm that the test step was actually
called. This also adds a duration to the test step - the test waits that
duration to respond. The integration tests should now confirm that the
test step was responded to because of this, and it lets us return errors
up to the point that the duration expires.

As part of the duration implementation, the test steps had to be
refactored to be better async citizens - they were using the std::fs
functions which blocked the executor, so this switches them to use the
async variants provided by fuchsia_fs.

Original-Fixed: 108753
Test: fx test blackout-target-test
Test: ffx storage blackout integration
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722125
Original-Revision: 63afb58c0a7e9d000fc094dd8bead59f29c72390
GitOrigin-RevId: 9083c83aa61bd9de30c56dcaacc2221c9a11fbd7
Change-Id: Ic2d6229fb31db6cbdee04a9a94e1cba21726034f