[roll] Roll fuchsia [assembly] Add set_source in AssemblyManifest
Sometimes we would like to move an AssemblyManifest, therefore the
source paths of the artifacts need to also change. This CL adds a
set_source function to facilitate this.

This is necessary for the AssemblyManifest to be used in ProductBundles
and for the paths in the AssemblyManifest to be adjusted during the
construction of ProductBundles.

Original-Bug: 107194
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718945
Original-Revision: 63e4254b64cce766d3211db25f2ecdcdf07ade1e
GitOrigin-RevId: 7878b8a70f4996958e4c4c4d90c1dbf9b6588575
Change-Id: I42ea8c6b3c8d8b7e13a696d12a9302090ab97ac2