[roll] Roll fuchsia [overnet] Don't force a global hoist()
Right now any process that includes the hoist library gets a global
hoist object introduced at process startup that is shared for the
lifetime of execution. This is especially problematic for tests, where
if they're run in the same process (as they are in any library/binary
not set to panic=abort) they'll all try to share the same hoist object
and even if they run serially might run into problems with that.

This change removes that default, instead making it possible to use
stored hoist objects. A global one *can* be set at process startup time
as well, but doesn't have to be. Eventually ffx will probably not set
the global hoist object to make testing easier. For now though, this has
every process that uses hoist other than tests set up a global one
(maintaining the status quo), and makes minimal changes necessary to
have most tests use a stack-allocated hoist.

As part of this, the Hoist object also becomes clone()able, allowing it
to be treated more as a passable handle than a singular global &'static
object and all that comes with that.

Future follow up changes will apply this mechanism to ffx and make it
more testable.

Original-Bug: 108692
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/721022
Original-Revision: 801d87fdbad404de7bab27f6ef587d59c7d4526b
GitOrigin-RevId: 26f9f08d5dd1eaefb53fce9ac8e2898229145b6b
Change-Id: Ib85e463ba5842af2602d5804648ced068651caed