[roll] Roll fuchsia [lowpan-ot-driver] Update backbone interface enable logic
Previously we needed to use a hack in the backbone code due to the lack
of support in netstack for us to differentiate wlan AP interface and
wlan client interface.

Now this support is available.

Update code to enable backbone interface to enable when the wlan client
interface is up. This covers the corner cases like ipv4 only and ipv6
only network, as well as the case where there may be no internet
availability.

Test:
(1) flash the device
(2) reboot the device, confirm the thread network is up
(3) using `wlan-dev` to connect device to a network with no internet
connectivity
(4) check the log to see if the backbone interface has been found, and
lowpan-ot-driver has been restarted to enable the backbone interface.

Original-Bug: 108782

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/721039
Original-Revision: ac9b9c7e9d3baa552f46ab08d3c81ad472317da5
GitOrigin-RevId: 72bd3490167fae60e87e3a2f63c5e8486de10eef
Change-Id: I1c174576f96c41f8977c02b5a99b397f09002812