[roll] Roll fuchsia [boards][suspend] Fix enable_suspend includes.
Define the enable-suspend-from-board group which includes the kernel
dependencies needed to enable suspend if the enable_suspend flag is set
to true. This group is then always included in the x64-common board.

This gets around the fact that the board is imported before the
enable_suspend flag is set.

Original-Fixed: 109080
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723029
Original-Revision: cc51ea01a7bf51b8dca4cd49a21fea8c19943c37
GitOrigin-RevId: 78e855bd5d856e3c371ce763bab91411072a5e2c
Change-Id: Iab70a83a67bf3315edecafed7c0ed45d592d0b9d