[roll] Roll fuchsia [netstack3] Drop `Default` trait for `Buffer`
the `Default` trait is only required because of the blanket impl of
`Takeable`, a Buffer is not required to be default constructible and
should only be constructed through `BufferProvider`. Fixed an incorrect
use of `Default::default` for transition from SYN-RCVD to FIN-WAIT-1.
Tests are still allowed to use `Default`. While I am here, I also
dropped `Buffer` as the super trait of `Residual` as it was never used.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718417
Original-Revision: e7a11f47906451a5fb4dbfbe51073242d0fc69d5
GitOrigin-RevId: 51b8634a9a3bb925308e2baf575dda9fc87d651a
Change-Id: I275c141f1c71fb3e02635d7065358781ecf04ed5