[roll] Roll fuchsia [kernel][vm] Correct anonymous page debug checks
When kAnonymousIsReclaimable is true anonymous pages, but not anonymous
zero fork pages, are placed in the Reclaim queues. This change updates
the Debug queries to correctly match this.

Since kAnonymousIsReclaimable is currently hard coded to false this
does not have a visible change that can be tested yet.

Original-Bug: 60238
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722183
Original-Revision: 9338d85692923b3d68ee7c0eb055be87ed916441
GitOrigin-RevId: c29b13201b490399b9daa9d403dccea44d99b04a
Change-Id: I1cc571301546d224f396a0fc665c712cc4e4d637