[roll] Roll fuchsia [recovery] Re-enable boot simulator test
This was disabled for the CFv2 migration. It depends on serial log
output which was implicitly emitted in CFv1. An extra change was
required to re-enable serial logging for CFv2.

This change also moves the key log line to the setup method of
AppAssistant, which is further along in app initialization than the
previous "recovery:started" message.

Original-Bug: b/241963587
Test: fx test -v recovery_simulator_boot_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719028
Original-Revision: ef8906f5596bf53e4c06e9478d7188f9f452b74c
GitOrigin-RevId: da040c24cec92431da0665423dd133bfddea2498
Change-Id: I2cd5ad72b195f37c20de1b5d6c2b76107f7c7301