[roll] Roll fuchsia [media/vnext] delete vnext prototype code and FIDL
This CL deletes the media/vnext prototype code and associated FIDL
definitions that are not otherwise in use. Given that the API design
is changing and the fact that this code uses HLCPP, this prototype
will not be useful in this form. Some of this code will reappear in
a different form as V2 work proceeds.

Test: deletion only, no new behavior
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723465
Original-Revision: f7a61a215495159b1a35aa5f9ddddaac94b2927a
GitOrigin-RevId: 8d026dabafbbd77689a3bc9b041e01a01b84ab07
Change-Id: I88c42a0eef0420e25d42a09c6783307b3c20bc3f