[roll] Roll fuchsia [netdevice-migration] Fix race in releasing netbufs
We are signaling to netdevice-core that a transaction is
completed and that a buffer is available before actually releasing
that buffer. This can result in netdevice-core sending a new
buffer before the buffer is returned to pool, resulting in
a false contract violation.

Resolve by signaling to netdevice-core only after releasing the buffer
back to the pool.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722967
Original-Revision: 9a0ef8d965d2d9f6a9cc0cbad1629a0f57052778
GitOrigin-RevId: 1d014aece396fc27189ae7c5693e0ad129a92c90
Change-Id: I2b0d4c58e26a284c2b85eb4c589261876ab051f2