[roll] Roll fuchsia [network/tests] Deflake WLAN AP / DHCP test
Currently test_wlan_ap_dhcp_server tests, among other things, that when
the DHCP server is running on a network, it will configure a host
endpoint that comes up on that network. It does this by waiting for an
interface to appear that has an address assigned within the subnet that
the DHCP server configures. The test repeats this process multiple
times, removing and re-adding the WLAN AP and host interfaces each time.

However, the logic to wait for address assignment waits for *any*
interface with an ID that is not the WLAN AP interface's ID. This means
that when the host interface takes a while to be cleaned up, the second
iteration of the test sometimes thinks it observes address acquisition
on the new interface when it is actually observing the already-acquired
address on the old interface. This means the test can begin teardown
when the DHCP server and client are still in the middle of
configuration, and this can cause a spurious ERROR log from the server
when an interface backing a socket it's using is cleaned up.

To address this, wait for the host interface to be removed after each
iteration of the test.

Original-Fixed: 108559
Multiply: netstack-management-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722638
Original-Revision: 31f08b6eaf999e9a7a255c880ead414575dd49ed
GitOrigin-RevId: cadbe1c02da57a564969d650129051c0eb3b782f
Change-Id: I8652bde0cbba8d214278d0e7cc31a13fa25e21c5