[roll] Roll fuchsia [icu][build] Adds third_party_icu_headers template
This is used for the ability to place the ICU library
into a directory other than the default //third_party/icu.

Some products need to use the ICU library from a different
directory, and this template makes the changes necessary
for that to be possible.

Original-Bug: 108544
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720284
Original-Revision: ca2ed71e317d8acfd02ebf8e218c0a237d4c89b8
GitOrigin-RevId: 36278772513a51df22ee5d8d0083798a181000bb
Change-Id: If37ad54e059c77eab51e77aca96dc4e81e66e829