[roll] Roll fuchsia [storage] change DELETED watch event name to "."
Change the name returned when the DELETED watch event is sent to "."
instead of the name of the directory.

For all other watch events, the name sent by the event doubles as the
relative path from the directory on which the watcher was set. This
brings the DELETED watch event in line with that behavior to make it
more consistent.

This cl also adds support for the DELETED watch event to fatfs and
minfs.

Original-Fixed: 76762
Test: fx test watcher-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722051
Original-Revision: 3b44021dd3db754b036c9abaa496d50b998c3eff
GitOrigin-RevId: 6352216afb30eab309914455e8ddb6303bc649da
Change-Id: Ie2db58e371e911466614a59a9c3dce06e9b12c0d