[roll] Roll fuchsia [pci] Migrate PCI from Banjo to FIDL under the hood
This changes the internal implementation of the ddk::Pci class to use
FIDL instead of Banjo. It also changes all users of the class to bind to
the FIDL PCI nodes in the device tree, and updates some instances of PCI
topological paths in the source code.

There are a few remaining PCI Banjo users that are not migrated by this
CL because they do not use ddk::Pci, notably iwlwifi, e1000, and
brcmfmac.

Test: booted Chromebook, NUC, and emulators
Original-Bug: 99914
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714013
Original-Revision: 424c30094198e8cf4a602b81e082fb071dcd07db
GitOrigin-RevId: 1401b5ea196a2bd9589399d07ea89bfc60cd919b
Change-Id: Icf8c8e42501ef59b518e9bd7960691e78cd9737c