[roll] Roll fuchsia [fidl-fuchsia-pkg-ext] Add serve_fidl_iterator_from_stream
Serves a FIDL iterator from a Stream of payloads. Useful for when the
payloads are generated concurrently with the serving of the iterator
(e.g. the fuchsia.pkg/NeededBlobs.GetMissingBlobs iterator will
be served concurrently with the handling of
fuchsia.pkg/NeededBlobs.OpenBlob when subpackages are implemented).

Original-Bug: 109245
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723537
Original-Revision: a271b469f1089d2e2893db2264f6baa5460214f0
GitOrigin-RevId: 0b30f5474b89d563684c4c051f84aa7e9f036679
Change-Id: I7d8a4e934a2b6323962125e68dc9944528d8691c