[roll] Roll fuchsia [testrunner] Add QEMU and home directory mounts
Adds a couple more mounts to keep existing host tests happy. The usage
of the current user's home directory is particularly hacky - we maintain
isolation by replacing it with a tmpfs in the sandbox, but the usage of
the /etc/passwd to find the name of this directory by a couple of host
tests is not ideal.

Original-Bug: b/211504655
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723422
Original-Revision: 4d6d8c5587daf310364e8c97e0dba787af3da779
GitOrigin-RevId: 8f2e8043811c09123740731eb44e340b08041caa
Change-Id: I67b455e6574771659de1d5b4fb7efff6e75b5f5c