[roll] Roll fuchsia [debugger] Use CommandContext in nouns.
Uses the new CommandContext object in the noun processing. This creates
a new ConsoleCommandContext just before calling into the noun system to
limit the scope of this patch and keep the rest of the system running.

The rest of the command processing will be converted in subsequent
passes. Then the CommandContext will be passed in rather than created
for the noun system by DispatchCommand.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723479
Original-Revision: 13a1709e67045495c9b56f31ffb82888a56d7471
GitOrigin-RevId: 941b1a515772b8d9db17966f1f5c784a73d2df5d
Change-Id: Icdf7c1c0c059d2cd46168b9bd82477e6c4d5fa50