[roll] Roll fuchsia [zither] Fix name utility inconsistency
zither.(Upper|Lower)CaseWithUnderscores() previously included the
library name as tokens, something zither.UpperCamelCase() does not. The
inclusion of the library name is rather C-specific (for want of a proper
library namespace), so this change removes the library name support in
the zither package and creates a wrapper with this support in the C one.

Original-Bug: 51002
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723782
Original-Revision: f023de76b0e27a8c719f449062647fbd5659f173
GitOrigin-RevId: eb73c8ecec640c1663ede3360867824e3ca48346
Change-Id: I82d341eff015655aa37386d78ed8bfa6dbd3f15d