[roll] Roll fuchsia [kernel] Log more when a process critical to root job dies
Any crash in a process that will bring the system down when it dies
is an important thing to log about in as much detail as possible.

This adds direct kernel logging about every exit from a critical
process, so that in all paths at least the retcode will be logged.
In the exception path, it adds full logging of exception and
register details for any exception fatal to a critical process.

Original-Bug: 108859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723373
Original-Revision: f53142df3d581067056196b952247ddf9d4bf090
GitOrigin-RevId: f6000e72cc2806e699879edb0ac861dbd9da6487
Change-Id: I20619204965ef4ed853cd4dca89732203458bfe7