[roll] Roll fuchsia [fshost] implement basic minfs/zxcrypt support
Add support for unsealing/formatting zxcrypt and mounting minfs on top
of that. For now we always use null keys for zxcrypt, which is fine for
the test, since that also hard-codes null keys, but the rewrite needs to
support getting the hardware keys as well. fxbug.dev/109067 tracks the
support we need for that.

Original-Bug: 87511
Original-Bug: 109067
Test: fx test fshost_integration_tests
Test: fx test fshost-rust-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723369
Original-Revision: 8daf658bba3658e96b1dbaf0f8a5c04ddf205303
GitOrigin-RevId: 6f683dd2830f74a8e3b58162230ac3e1b23db164
Change-Id: I3c376542ffdb29e12c76d35f939d20bfbaf72ab4