[roll] Roll fuchsia [fidl][example-tester] Create example-tester lib
The example-tester is a lightweight testing framework for testing FIDL
canonical examples. The key usefulness of the framework is that test
assertions may be performed on log output only. This is useful because
the primary purpose of canonical examples is not to make for good,
complete, robust tests, but rather to serve as terse and readable code
samples. In this light, a test framework that encourages logging of
important code checkpoints, but does not produce any other output that
would clutter up example files, is desirable

Test: fx test example_tester -r --count 100
Original-Bug: 107613
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694933
Original-Revision: 5c30058d83892704a87cfab9d2c12a3a682cb2e1
GitOrigin-RevId: ffd414a68e88795504a2a62688ccc6844a7b7674
Change-Id: Ie89b0efdeddca862dfe04c3be2d05547e5ec7a30