[roll] Roll fuchsia [ffx][daemon] Graceful exit always waits a bit
When adding the mechanism where the daemon gracefully shuts down when
the socket file is deleted, we accidentally made it so that this
bypasses the wait from a graceful shutdown message intended to allow
existing messages to finish processing (including and especially the
reply to the quit() message), because it would delete the file
immediately and then the delete handler wouldn't wait.

This moves the delay logic into the main loop itself, making it so that
*all* paths through the quit logic go through the same process,
including the brief wait.

Original-Fixed: 109242
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724084
Original-Revision: 2eb9b9636eb6de6c895e12a849eedeb79dd5f697
GitOrigin-RevId: 772861e42ff9e3892a1b601e094686468241f679
Change-Id: Iab8b250af8370e01cf1bacdf2017a5804f4a14e9