[roll] Roll fuchsia [settings] Add custom watch function for audio
Because we increment the modification counters stored in AudioInfo for
every change, clients are notified for every change, even if it is
duplicated. We need to keep the modification counters for earcons to
properly play on voice change, so a change function is instead added to
the Watch fidl handler to only respect stream changes.

Test: fx test setui_service_tests, audio_integration_tests
Multiply: audio_integration_tests
Original-Fixed: 107775

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722022
Original-Revision: 29966429a7261edad3523d09ec6a43d986c9b142
GitOrigin-RevId: 5ad07acc303fbb0959361368c429dec907e6cffa
Change-Id: I74c90ec78c0861d7a6993b5467a0a157e3535af3