[roll] Roll fuchsia [Settings] Reland "Convert Display to use jobs"
This relands fxr/707813 with a fix for a loop where a client that
watched the light sensor with a delta of 0 would continually receive
updates on their hanging get. See the bug for details on the investigation.

The diff between the original CL and this fix can be found at fxr/722965.

Converts the `Display` settings to use jobs. Also removes an unused
hanging get function.

Original-Fixed: 79041
Test: fx test setui_service_tests
      manually verified through device logs that there's no loop of
      light sensor watches
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722963
Original-Revision: cc34278620d67aee754c07c45b39075ce15d85a1
GitOrigin-RevId: e30a71b54a511732ae948fe17e8c5572249518e1
Change-Id: I595eb74f8fabb0634f448ea52d3036b718a17473