[roll] Roll fuchsia [fidl][cpp] rename Unknown(Interaction->Method).
Rename FIDL Status::UnknownInteraction to Status::UnknownMethod.

This status is only used in clients to identify when a method call made
by the client was unknown to the server. Unknown events and unknown
methods on the server are handled using a handler interface rather than
an error status.

This renaming is intended to make naming more consistent across
different parts of unknown interaction handling. The handler interfaces
are named UnknownMethodHandler and UnknownEventHandler, so naming the
corresponding error UnknownMethod is more consistent.

This is part of a set of changes which will make names more consistent
across all languages.

Original-Bug: 88366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722632
Original-Revision: ac01bc297330f509a9dc3e9873f202e08214bb8f
GitOrigin-RevId: 8e74fcc8ece9fed5075944e1795f36e85ea7db43
Change-Id: Iab08445aa79f499718d0b39992ba3a5ea504a673