[roll] Roll fuchsia [build] Have assembled_system() emit image metadata
Currently, all relevant image metadata is specified outside of
assembled_system(). This is largely due to the fact that image metadata
is quite a mess and contains a lot of information that is out-of-band to
the template. This change moves the canonical image metadata into the
template, keeping the wonkiness in the user-defined domain with a new
`image_metadata_overrides` parameter.

Original-Bug: 81569
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723490
Original-Revision: 813f55ed453a5e97bbe86450cc56ad4b2799dc88
GitOrigin-RevId: 66be729f9e3547022b66e2aafdedf5fcf9cc300a
Change-Id: I7ccde18cd83b8e0fb80d887fd711d1b54eadc0e2