[roll] Roll fuchsia [kazoo] Replace golden file testing scheme
The previous way concatenated every kazoo output into a single, very
large text file and mediated golden diffing through that with a bespoke
scheme. The sheer size of this file and its presentation as raw text
made it rather difficult to grok what the kazoo outputs actually were -
and updates had to be manual.

In this change we remedy that by replacing that infrastructure with now
canonical `golden_files()`, which allows us to separate out the files
(presenting them as proper source code), remove the bespoke logic, and
update them via a simple build or more directly with
`fx check-goldens`*. We check in the goldens under
//zircon/vdso/goldens. (Without further reorganization, this
is the most convenient place to put them as a target in
//zircon/vdso/BUILD.gn is what currently generates most of the files in
question.)

*In this case the minimal way to update them is `fx check-goldens vdso`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724314
Original-Revision: ae28f60802ec79a0d68f0ea655de34e8f35ba1a7
GitOrigin-RevId: 26cba8db9787e4a59bab1850104a632586b35fd2
Change-Id: I547ab9c1ade802ada24650fd7114bc43d87b65c8