[roll] Roll fuchsia [icu][build] Use third_party_icu_headers
This is used for the ability to re-target the use of
the ICU library, which is required for some products.

We need to replace the use of third_party/icu with
a variable root directory to be able to vary ICU use,
both in BUILD.gn files, and in C++ files where
the #include directives refer to header files by
their path.

Original-Bug: 108544
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719663
Original-Revision: c3f0f3f25b9e0d26c7039274c93f0c7abd33e848
GitOrigin-RevId: 78733087cd358250c4dcd661f905f0bf82bbb18e
Change-Id: I5fbcf5ebc227daf9d4d1db87f70a61e78de942b2