[roll] Roll fuchsia [icu][build] Use third_party_icu_headers
This is used for the ability to re-target the use of
the ICU library, which is required for some products.

We need to replace the use of third_party/icu with
a variable root directory to be able to vary ICU use,
both in BUILD.gn files, and in C++ files where
the #include directives refer to header files by
their path.

Original-Bug: 108544
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720287
Original-Revision: cee6bddb91dddaf8f74ebba104c4ef97c2bfcdea
GitOrigin-RevId: 4e41244fbbb12af33f8a98ae69267719090860b7
Change-Id: I5ab234543ce7191f3ff0c9e85d9a8084f9e6b99b