[roll] Roll fuchsia [icu][build] Use third_party_icu_headers
This is used for the ability to re-target the use of
the ICU library, which is required for some products.

We need to replace the use of third_party/icu with
a variable root directory to be able to vary ICU use,
both in BUILD.gn files, and in C++ files where
the #include directives refer to header files by
their path.

Original-Bug: 108544
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720286
Original-Revision: f903f840862e55e5d200b92583d43c1ea918b44e
GitOrigin-RevId: 046f31a6d1d926ef5322ed4b8651097fbcb58ae4
Change-Id: I5533c16624066c095d6f45299a6f2c3daa65d394