[roll] Roll fuchsia [kernel][sched] Remove `period` from deadline profiles.
Remove the `period` field from the kernel implementation of deadline
profiles.  Having a period separate from the profile's relative
deadline was a feature which was not really being used, and it was
making the generalization of deadline profile inheritance difficult.

So for now, remove the concept of deadline period from the kernel's
implementation of deadline profiles.  After this, we can either remove
the period from the profile definition entirely, or we can
re-introduce it once we figure out how to properly incorporate it into
deadline inheritance.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658372
Original-Revision: eed5c7381c7e469f66f31bd724a353f3823ccbb9
GitOrigin-RevId: 8d16a5cc4c43cced543290e8860b7c3211b52149
Change-Id: I339df3a7537018ba6580b3c5cbcba5ff8ee0b081