[roll] Roll fuchsia [usb][xhci] Do not attempt to track event TRBs in flight.
Before this CL, each event ring (one per interrupter) was attempting
to track the expected number of events which are "in flight" at any
given point in time, and dynamically grow the event ring whenever that
number was larger than the number of TRBs currently in the event ring.

There were several issues here, but the primary one is that there is
really no such thing as "tracking the number of event TRBs in flight".
An interrupter in the host controller is allowed to produce an event
TRB asynchronously (IOW - any time it wants to, solicited or not).

There is a mechanism in the system (the PCS/CCS bits) which give a way
to software to detect that it has failed to keep up with processing
the events produced by an interrupter, but there is no way actually
control the number of TRBs the interrupter will post to its event
ring.

So, there is really no good "worst case" accounting we can do here.
This is a game of "keeping up with the" interrupter, not actually
budgeting for the worst case.  If the interrupter ever produces TRBs
faster than we can consume them, then we lose the game.  How many do
we need?  It will depend on how quickly we can actually respond to
interrupts.

At any rate; this whole system was causing problems.  Some of them
included:

1) Accounting was not being done properly for asynchronous events,
   meaning that at startup we would receive a number of plug events
   causing us to decrement our "in-flight" counter and causing it to
   underflow (meaning we thought there were ~2^64 TRBs in flight).
2) There was some issue where immediately after we started to become
   operational, we thought that we had over 2000 event TRBs
   "in-flight", causing us to grow the event ring many times during
   startup.
3) There was some "leak" in this accounting which was causing us to
   leak TRBs at a rate of 6 per minute on Atlas.  This caused the ring
   to keep growing until it saturated the event ring segment table.
   At this point, a series of unfortunate events was causing...
3a) Code to allocate a VMO, map it, and then pin it.  (heavy ops)
3b) Code to then realize after the fact that the segment table was
    full.
3c) Code to then discard the VMO it had just created (leaking the VMO
    and the mapping because it didn't realize that it had to unmap the
    VMO, not just close the handle).
3d) Finally, the code ended up ignoring the error reported by the
    event ring class (failed to grow the table), meaning that the
    system just kept operating.

So, after about 16 hours of leaking accounting, the system started to
flail, creating, mapping, pinning, unpinning, and leaking VMOs
rapidly.  All of these ops were being done on very time critical
threads, and the more mappings which got leaked, the more expensive
the heavy mapping operation became.

So; this CL does away with a lot of that.  It:

1) Statically allocates the size of the event ring, attempting to have
   space for 2048 events per ring.
2) Adds inspect properties to allow us to monitor the maximum number of event
   TRBs which are processed during a single IRQ.  This is a
   conservative proxy for the worst case RD -> WR pointer distance at the
   time of an IRQ, and can help us to understand if we have the proper
   tuning for the size of the ring. So far, I have not seen a number
   greater than 401 on atlas with both USB audio active, as well as a
   heavy ethernet load, but I have have not stressed the system much
   harder than that.
3) Limits the number of interrupters we instantiate to 2.  We never
   actually see to make use of more than 1 right now, so there was
   little point in allocating 8 on Atlas.

That said, this is just a quick fix.  There are _many_ more
improvements and cleanups which can be made here.  In no particular
order:

++ There is no reason for the complexity involved in the current event
   ring TRB management code if we are not going to dynamically grow or
   shrink it.  Eliminating this complexity would allow us to perform a
   single VMO alloc and map, meaning that the virtual erdp pointer
   would traverse a virtually contiguous ring, even if the underlying
   pages were not contiguous (eliminating a bunch of compled code
   which is trying to get to the next segment in the ring since it was
   not mapped virtually contiguously).  This should also save TLB
   overhead due to the single mapping instead of multiple mappings.
++ Improve the CCS/PCS overflow handling code.  Right now, it
   basically does nothing at all.  If we ever do encounter an
   overflow, we really do want to log an error to bring it to people's
   attention so that they know that the tuning parameters may need to
   be adjusted.
++ Make the tuning parameters a system configurable number (think;
   driver parameters in the windows registry), so that various targets
   can be tuned independently.
++ Actually spend some time focusing on load balancing the
   interrupters.  Currently, we basically only use one, and we can
   probably do better if needed.

Original-Bug: 102058
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722962
Original-Revision: b6d08a00d16ccc3c342b7f78de7404508051a8d8
GitOrigin-RevId: da848a9194991d06f4131f4c7413de8014b303a3
Change-Id: Iaf7bee6f5bf22708a47b81ed8976d0bee05927a4