[roll] Roll fuchsia [docs] Add documentation of the IfThisThenThat checker
IfThisThenThat is a presubmit check that runs automatically. However, it
requires the user to understand that it exists as it operates on special
comments that need to be inserted into the code (most presubmit checks just operate
independently).

Because IFTTT requires some teaching, it needs to be documented. As
there is no current page addressing presubmit checks, this change adds
on and places the IFTTT instructions within.

NO_IFFT=This is example code, not actually IfThisThenThat directives.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724597
Original-Revision: 29da67a9447ca333145b86b1bef28b0d8b6d4dcc
GitOrigin-RevId: 8c8dad59a6ae7c485f5eef849a902ab86dd691fc
Change-Id: Ib91fac5e094446d5bfbf7b10d1b350442c93d75b