[roll] Roll fuchsia [fidl][cpp] Remove deadline support in sync calls
If the deadline elapses, we do not close the channel and the response
from the server would not be recognized. Because deadlines are not used,
it is easier to remove them now while we discuss what's the best
strategy for sync deadlines.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724825
Original-Revision: da7b96a1edf8aa9e66eb93dceee9153ac6ad829c
GitOrigin-RevId: 3d67135895670e35584cacbde3c51a76c299298d
Change-Id: I12d97edc0006c9c18d6d3d766ea52cabfc546125