[roll] Roll fuchsia [assembly_input_groups] New bundle
Introduce assembly input groups, which represent a collection
of artifacts that correspond to partial input to a system image.

A group is generated for each attribute supported by the input groups,
each of this groups represents the list of labels that are to
be used in the assembled_system() template.

An extra field, is the set of core realm shards, this are used
to generate a core realm that should be embedded in the base
package set.

Manually tested by locally replacing the usage in fuchsia/BUILD.gn
to point towards this label instead of doing lookup.

Test: fx build
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708584
Original-Revision: f6588402c5f2bb9f49393b7f3cd56f724bc94a6a
GitOrigin-RevId: da56e328c65fc7a408c9241aca73d7bbd1eb1bd5
Change-Id: Ic888518751a245cf9cc855c8d4b920037fbd4f4e