[roll] Roll fuchsia [debug_data] Don't schedule processing task if processing is done
This removes one potential redundant write that could occur. This also
fixes a flake where writing to the summary could be attempted after the
directory is torn down.

The flake could occur depending on the order in which AddData,
DataFinished, TakeMapContents, and FinishProcessing methods are called.
AddData and FinishProcessing are called from thread 1, TakeMapContents
and DataFinished are called from thread 2.
In case the order is AddData -> TakeMapContents -> FinishProcessing ->
DataFinished, the dirty bit is still set when the processing task
finishes. This causes thread 2 to dispatch the data processing task
again. This could trigger attempting to write to the summary file again.
With the recent change to support isolated debug data directories, at
the time this write is attempted the directory may by in the process of
being torn down (as it is managed by a separate component).

Multiply: test_manager_test
Original-Fixed: 107144

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/725369
Original-Revision: 840f6793c1a6362c1dc0d3ffdddab930452de07d
GitOrigin-RevId: c4abc48e38ca1d1e6bcaa9f92623f081dadd2734
Change-Id: Ic6cc76242399b8e0c00d07106728f41dc8396a07