[roll] Roll fuchsia [icu][build] Use third_party_icu_headers
This is used for the ability to re-target the use of
the ICU library, which is required for some products.

We need to replace the use of third_party/icu with
a variable root directory to be able to vary ICU use,
both in BUILD.gn files, and in C++ files where
the #include directives refer to header files by
their path.

Original-Bug: 108544
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720285
Original-Revision: e91710ec7d4fc42b00f95b181550b37fb71105f9
GitOrigin-RevId: dc5e6d0335ea4c5d870e39b4880645403f489e89
Change-Id: I284e939032c42b1dd6a5b6f6042baf29dc75545b