[roll] Roll infra/recipes [reported_step][build] Upload failure summary
Upload the failure summary of a build as part of the
artifacts of a reported step. It was require to refactor
the code of reported step to allow to append artifacts generated
by later steps than the one that wants to be uploaded.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/724704
Original-Revision: 2b28891a3ec4b587f67557c5de088b2637e588df
GitOrigin-RevId: d8f786702f6218dcdc0f0fb8be0bb682e1f22011
Change-Id: I66f0d531528b6113dbb66f95d7dbb05d82fee08f