[roll] Roll fuchsia [dhcpv6] Add collected_advertise to all states
The collected Advertises need to be tracked also in Assigned and
Renewing states because it may be necessary to try another server
when Renewebing/Rebinding if no usable lease is found.

Also assert the value of `collected_advertise` in tests wherever it is
convenient to do so (which basically means whenever it is supposed to be
empty, or the want value is known due to being directly constructed by
the test or obtained by inspecting a previous `ClientState`).

Mechanical changes only, no change in behavior.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723728
Original-Revision: 58939c8afa2b41113a468f0c761595aa4766e9a0
GitOrigin-RevId: 0fbd86ca07062738c793018b0e19de2b2dc993d1
Change-Id: Iebd780fe5dfa3773323f7a0ec7c849a2a1603af2