[roll] Roll fuchsia [identity] Persist AccountMetadata in AccountManager
We were ignoring the account metadata which was sent (or not) as part of
provision_new_account api. This change enables AM to start persisting it
as part of the stored_account_list.
This will most likely get moved to AccountHandler in the future.

Original-Fixed: 101664
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724409
Original-Revision: 29ed94c5322266dd5a34b4c8e6b01231dbd62646
GitOrigin-RevId: 7b7908d41309fb487f6040a518842c90c8bf3e39
Change-Id: I75c1e5a603aa3b6b19c115197366d3b6464555b4