[roll] Roll fuchsia [fidl][fidlc] Use one FIDL for error doc and test
Tests for specific errors are now able to have their test case FIDL
inputs factored out into separate files, thereby allowing that input to
be used both as a test case and as the includecode for the markdown
documentation of the error in question. This creates a single source of
truth for both error documentation and the tests it documents.

Test: fx test host_x64/fidlc-test
Original-Bug: 108248
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718043
Original-Revision: 9d9e829b51e1de701fd7535a3bb985271d9ada10
GitOrigin-RevId: 96cc8ca27d4fb3d5258a4ad210a9d15a90b1f479
Change-Id: Ifd5546d332e76180ca20a6be34464f2a6997d5da