[roll] Roll fuchsia [fdf] Create DFv2-specific workstation_eng groups
Create the GN product `workstation_eng_dfv2`. This is identical to
`workstation_eng` except it doesn't include tests that are failing when
DFv2 is enabled. Also create the GN group
`bundles/buildbot:workstation_dfv2` which is identical to
`bundles/buildbot:workstation` but it does not include tests that are
failing when DFv2 is enabled. Also separate buildbot workstation tests
into two groups: one for tests that are included in host labels and one
for tests that do not.

Existing builder bots with DFv2 and workstation_eng will use these
counterparts in a future CL. This is in order to get builder bots that
are running DFv2 passing and monitored by the Fuchsia build gardeners in
order to prevent regressions in DFv2.

Original-Bug: 105115
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718666
Original-Revision: f82a4d4e973e9c27d2026a6fa91bd44a5ad388b7
GitOrigin-RevId: 8cec3300f87a115da3661b3cb854fbf6110cd073
Change-Id: If96b00d6bf554a907cf877947e90a087e845f483