[roll] Roll fuchsia [cm] Migrate destruction tests to events_v2
This CL migrates the destruction tests to use events V2 per RFC-0121. The tests
are refactored to deal with limitations of subscribing to events from within
a nested component_manager by not expecting to receive events in a particular
order or a particular set of events. Instead, it tests for a subset because
we may receive more events than we're expecting because there is no way
to subscribe to only a certain set of events.

Run-All-Tests: true

Original-Bug: 81980
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722524
Original-Revision: 37380676b81d7b8bbf15b5408d50d57c708e1262
GitOrigin-RevId: 676ca737d2ac15b7811bc609d64722d06f0d1bcc
Change-Id: I2b23e3e755e37c6ede3487e876d30b8e601892a2