[roll] Roll fuchsia [compat] Wait for unbind to be replied to before calling release
Right now we call release under the assumption unbind completed right
away. In reality, unbind is my be compeleted asynchronously. This may
deadlock as we are blocking the dispatcher from running. We detect that
via a 10 second timeout. This is not ideal, but crashing in this case is
probably better than calling release incorrectly.

The proper fix for this situation will require the prepare_stop hook to
be implemented as described in go/tq-dfv2-dispatcher-shutdown. This will
allow us to trigger unbind durring prepare stop and release during stop.

Original-Bug: 97457
Original-Bug: 92196
Test: fx test compat_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724987
Original-Revision: ee7d776d402d386b782e94c72284671ed119df18
GitOrigin-RevId: 2a32bfe58881a8300a35ec086748d34c409fcf3a
Change-Id: Idfe8f0c23ab126580117a52a153998019bc3e91a