[roll] Roll fuchsia [ffx][daemon] Use build-dir for socket in-tree
Instead of using the global socket, use a per-build-dir daemon socket
path when using ffx in-tree. This prevents frequent fights over the
daemon socket when jumping between in and out of tree use cases, as well
as between different build outputs.

Also fixed a bunch of error paths involved to be clearer and/or do more
sensible things and not present as BUG: when they happen.

Original-Fixed: 109129
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723317
Original-Revision: e54c600d40072207da6010074d2801b8a4852190
GitOrigin-RevId: 0c6a368b9049d6eb324c65e345474797a42031ab
Change-Id: Idaa332b723ce989d93ab970e09cf477cfe161d00