[roll] Roll fuchsia [kernel][vm] Remove lock from content size field
This change removes the lock from the content size field in favor of an
atomic. This is in preparation for a following change that allows for
content-expanding writes to mutate the content size as progress is made
without needing to hold the `ContentSizeManager`'s lock.

As mentioned in code comments, removing this lock is logically sound
because the queues in `ContentSizeManager` uphold the invariant that
only one content size mutating operation can occur at any given time.

Original-Bug: 108771
Original-Bug: 107119
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724649
Original-Revision: dc28ac4dc9e6edcb56f13381592518f7f69032d6
GitOrigin-RevId: 98be6f412664f09fb109c789d0d6e0531a58e16a
Change-Id: I0a6f2857ac43aada4bb73b84e3ccb4d51725381e