[roll] Roll fuchsia [audio] Fix bug in Format::ToFidl()
When a FIDL struct contains a field of type union, FIDL requires the
field to contain a value since unions are "not nullable". This manifests
when you try to encode the struct, such as when calling or responding to
a FIDL method.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/725341
Original-Revision: 886b177305cc4afa6fee75e9c68da0499d7a41c1
GitOrigin-RevId: 13960d720422a450200172b037c4440076bb810e
Change-Id: Ice622215206e2dc209d9f11bd0b2e6ae90a6775e