[roll] Roll fuchsia [driver_host] Fix power state map indexing
An off-by-1 error for indexing into the system power state map seems to
have lead to certain code paths being optimized away. Specifically in
LTO builds, the `DEVICE_SUSPEND_FLAG_REBOOT_KERNEL_INITIATED` switch
case in `device_get_dev_power_state_from_mapping` was not being executed
as expected. Presumably this switch path was optimized away because the
result it produced is only used to index into an array, which would be
invalid in this case.

Since the SystemPowerState is a FIDL enum which is a 1-based index, we
should actually be using `sys_state - 1` as the index into the system
power state map.

Ultimately, it doesn't seem like the value from the map is used
anywhere, which is probably why this went undetected for so long.

Test: fx test driver-host-tests
Test: fx test host_x64/obj/src/tests/oom/oom_tests.sh
Original-Fixed: 105795
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723481
Original-Revision: ca16157034d69980f3951730baa5ea48d312ac2d
GitOrigin-RevId: 847befa90370d7ca87286a74f394dca7266bc423
Change-Id: I70f76e1029babd6fd5ad91f620f8657ed77d9949