[roll] Roll fuchsia [ffx] Extend `AddTarget` to return more info.
This extends `AddTarget` to return more info about errors. For example,
there may be an ABI mismatch, but it is important that the user know
what the ABI mismatch is (the local vs the remote ABI). This sends a
struct to the user when there is an error instead of extending an error
struct. Without this change, if there is an error that has additional information
included with it (like a specific ABI compatibility), the added target will be
automatically removed if requested to do so in the config.

Current behavior should be unchanged, as the return log of errors is not
intended to be parsed, simply relayed to the user.

Original-Fixed: 107759

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722124
Original-Revision: f22fca203eebfb883940bf4d2d20fb02686375a7
GitOrigin-RevId: e750b279bb8120f7e1bea402b00a454560faeb58
Change-Id: I078a9148316a69c05b2dd3b9c14aa0a8fbd2e516