[roll] Roll fuchsia [ffx][fuzz] Stop does not need to connect
This CL adjusts `ffx fuzz stop` to avoid trying to first execute an
`ffx fuzz attach` command. The latter results in a
`fuchsia.fuzzer.Manager/Connect` FIDL request, which is pointless when
the former is trying to send a `fuchsia.fuzzer.Manager/Stop` request.

Separating these requires that the `StopSubcommand` not be derived from
a `StopShellCommand`, and provide its own optional URL as an argument.
If the URL is not provided, `stop` will still use the one from a
previous interactive `attach` command, if present.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723494
Original-Revision: 243e8c4a6101e8fcb06a2b385dfbb139b6602abf
GitOrigin-RevId: 91946c6f3a3d16eac49601680aaf71e346b22ab5
Change-Id: I1da85312e989876fef0ac63c5b27274b147eb7ca