[roll] Roll fuchsia [iwlwifi] pass all multicast packets to driver
Currently we only ask the firmware to pass 3 fixed multicast
addresses to the driver. However, it is not enough for IPv6.

So the new code asks the firmware to pass all multicast addresses
to the driver so that the driver can forward the IPv6 multicast
packets to the netstack.

Original-Bug: 108681, 100129, 99200

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724077
Original-Revision: 002b6be1f5b836e347c1180a3266be4be361491f
GitOrigin-RevId: 2bdebbf742ebab94157707f7422ce6bd00a16458
Change-Id: I903947f563bb65b101346d83772f4dff4d960de7