[roll] Roll fuchsia [kernel][sched] Cleanup first thread behavior
Currently, the first threads created for each CPU are something of a
special case.  When they are initially created, the scheduler is not
actually completely running yet.  Threads are assigned a weight, but
they don't actually participate in the scheduler's bookkeeping.  IOW,
they don't do things like:

1) Contribute to the total fair weight of the scheduler
2) Contribute to the count of runnable fair threads
3) Contribute to the estimated runtime of the queue

This said, they _do_ actually do things like attempt to obtain locks
(for example, the heap lock, or the debug log lock) and can block as a
result.  This blocking can result in profile inheritance and
propagation, and that is where things start to become a problem.

Some of the asserted invariants of the scheduler bookkeeping can
become violated during this period of time, if there is contention for
one of these locks.  While this is not currently very obvious, it is
becoming more of a problem as I work through the work to generalize
profile inheritance to support deadline inheritance.

So, as a work around, change things just a bit in order to have these
threads participate more in the bookkeeping.  Add a couple of special
case hooks to the scheduler which allows these threads to participate
in the bookkeeping initially, and then drop out late on when they are
done.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470829
Original-Revision: 99faf85532cd159cd33dfb5d6e0a3647227468c6
GitOrigin-RevId: d6fb387603551237b27fff5f8f9b4d0afe10caf0
Change-Id: I8cd176827c00df025d90a297aee16b97841d3206