[roll] Roll fuchsia [build] Remove export-compile-commands switch
The toplevel .gn file now globally exports a compilation database for
the //:default target. It is no longer necessary to pass a switch to GN
to make this happen, and the uses of this switch have been removed.

Remove the zircon_comdb_filter variable. This is never used.

Original-Bug: 109452

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/725327
Original-Revision: 13edd57ac5e88a05e4ac4179962f0667bb2c8ad0
GitOrigin-RevId: 867a492e4faedd7f2d8bad40ed10739cfc980b8d
Change-Id: Ie927a1820511b2e82294fbbc78b944c7fc188784