[roll] Roll fuchsia [wlan][policy] add logs for removing saved network
Add logging for the case where an API call comes in to remove a saved
network and the SSID matches but the security type doesn't match. Log
the security type of the network that matches. This will help debugging
cases where the network is not correctly specified to be removed.

This change also gets rid of the key from the map of saved networks if
the only saved network is removed.

Test: manually tested logs
Original-Bug: 105235
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678945
Original-Revision: 45e891e9406660d95556fad75ca01d59e4b70f13
GitOrigin-RevId: 2d8c3bb58f17831f86868570bca901f3f1ab832d
Change-Id: I53c3e06bf7ca37f00c8ec8f4aa3611588f96168e