[roll] Roll third_party/pigweed pw_sync: add class InlineBorrowable
InlineBorrowable owns an object and a mutex that guards the object.
Both the object and mutex are constructed in-place, and the object can
only be accessed by calling acquire() or one of the acquire_*() methods.

InlineBorrowable should be used whenever an object needs to be guarded
for its entire lifetime by a single mutex. InlineBorrowable extends
Borrowable, so it can be passed to existing APIs by reference.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/103180
Original-Revision: 67e71ad7f7a3e8dfc63799d8ff241cd0a6daf9ee
GitOrigin-RevId: 8f79cb039bdbeb7bd6ddfd573af1bfbb2b6b09e6
Change-Id: I25e22fb74e50ba2010a443069592cf9d08137ac4