[roll] Roll infra/recipes [build] Improve error msg when fint crashes
If fint fails in a such a way that it fails to produce output artifacts,
the build currently fails with a confusing message about the artifacts
not existing, when the real error is that fint failed (e.g.
ci.chromium.org/b/8802589323816151345).

Now the build will instead emit an error saying that fint failed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/727022
Original-Revision: 60cca8f27e78ccffbb866dc179554f33822fff2f
GitOrigin-RevId: 184dab957fc1ef84340e67eadb09cd0cb9d6d7dd
Change-Id: I067077759560947a5290945802173208135efa57