[roll] Roll fuchsia [testgen][rust] Rename lib.rs -> testgen.rs
lib.rs is a module name that is the entry point for rust libraries, and
it has become a warning to name a module 'lib' in new toolchains.

Rename the generated module to 'testgen' which makes it clear where the
code came from and hints that it will be (eventually) refreshed.

Original-Bug: 109622
Test: refactor; fx build && CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726543
Original-Revision: 5e16cb8430eadf5bcda3c33a2a2085d8feb5ed6c
GitOrigin-RevId: 35bf327489f907dac5024c071e96e0c953cdb718
Change-Id: I07abe7fc727bd17807237734fd1988890280a5d7