[roll] Roll fuchsia [tools][fuzz] Allow stop_instance to run with older handles
If a handle is from an older version that had fewer fields, we can still
do a best-effort job of cleaning up instead of declaring the handle
invalid.

This scenario occurs when a ClusterFuzz build with a newer version of
undercoat needs to clean up after a previously-running build that had an
earlier version of undercoat with a different handle format. This won't
work for all possible version mismatches, but is meant to be a simple
stop-gap solution.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724823
Original-Revision: feaf3fbab827b8e57ebd3706bd8c90bda0c19f8c
GitOrigin-RevId: d6f1a10c946728162aa78e1eb2cbbbfa5335dd00
Change-Id: I5f6b49f282d62c14f7874a15fa421549b7a499f8