[roll] Roll fuchsia [check-licenses] Move project filtering to cmd package.
Currently, check-licenses calls the gn executable at the end of the
execution, in order to filter out the projects that aren't in the
current build.

This works, but makes it difficult to change the behavior of the tool in
some situations. E.g. in the future, if we know that a given project
hasn't changed from the previous run, we could potentially skip
processing that project entirely. But we need to do that analysis up
front.

This CL moves all of the filtering logic to the cmd package. In the
future, the cmd package will also do some preprocessing to recognize
areas of the tree that haven't changed recently, and skip processing
those parts of the tree.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726502
Original-Revision: 4eaee0abf8cdc7331ba3504b67b16ce307129a1a
GitOrigin-RevId: c607b30a3fb11572d3e6ebd075621f952c383646
Change-Id: I7c3e5afe86a8c7668b7673b8957b1de383ec672f