[roll] Roll fuchsia [reland][virtualization] Add integration test for container startup.
When this first landed it caused the sherlock-debug builders to run out
of space. Since we don't run this test on ARM we just need to make sure
we don't include the container images on ARM either to avoid this
problem.

Add a new type of enclosed guest: 'TerminaContainerEnclosedGuest'. This
extends the existing TerminaEnclosedGuest to complete the container
startup. Container images are pulled from a local file on a virtio-block
device instead of the internet.

In terms of performance, running the container startup takes ~1 minute
to complete so test that use the Termina container will be implicitly
longer. For now we just have a simple test to verify that the container
has successfully been created.

Original-Bug: 109706
Multiply: termina_container_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727317
Original-Revision: 5be23427b94ed3bf6331010ac261efcf83e690ad
GitOrigin-RevId: 78bbe14d5f2bd8a7ddce6136783433863ecfa89a
Change-Id: Ie3d60d2562afa3c48f74ea8b5180a1aa4be2803a