[roll] Roll experiences [gazelle] Tidy up the gazelle prototype.
This splits the previous `WindowManager` into two layers: `View` and
`Manager`. The `View` doesn't do anything smart at this point: it just
interacts with Flatland directly.

The `Manager` has policy in it (e.g., as soon as the child for a new
window is attached, transfer focus to it). It also handles queuing up
future work and executing it in the background.

This doesn't change any functionality, so it's still pretty much a
toy, but I think (hope) this'll be nicer to work with.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/727025
Original-Revision: efe4c2064f7cf8d31a6ee2ddd88680a4251b77aa
GitOrigin-RevId: 5f9d48f9f90d4fa4d1abdd64fe08df83b51407d8
Change-Id: Ia0f1e4d7003336d554a52e8eaa301d5403c5ff67