[roll] Roll fuchsia [vaapi] Don't always reconfig output buffers
When a mid-stream configuration is requested by the decoder check to see
if the current output buffers are large enough to hold the new size. If
so there isn't a need to reallocate buffers from sysmem. Instead just
send the new constraints to the client and keep decoding. Also added
some checks so that we make sure the new size is supported by the
underlying hardware so we don't get a nasty libva error.

Original-Fixed: 102737

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/696702
Original-Revision: 675e232a0f13478fd7e43cc1e9366290f5cb1fc7
GitOrigin-RevId: 698f7ee67dca45bcfc28a04e8bf8207c790da99d
Change-Id: Ibaebe4b19fd91d4122110086314006653c1e8be6