[roll] Roll fuchsia [fshost] Use hermetic crypt service in fixture
Previously, the same crypt service was used across different tests which
led to errors when creating another test fixture that formats fxfs. This
was guarded against with an atomic bool, but this is not sufficient as
the test realm may be re-used across test runs.

Test: fx test //src/storage/fshost-rust/tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/725267
Original-Revision: 00db7f8fb9936be22f979af6cfbeb0a333aef0b5
GitOrigin-RevId: 3250ac233d6f8ef3b73c54b2ce88bb50e3b1007c
Change-Id: I6c1423f1ef3b205997958c73fe28d3e286a8db94