[roll] Roll infra/recipes [reported_step][build] Upload failure reason
Upload failure reason as part of the artifacts of the
test results. This required changes in the script that
uploads to resultdb to be able to take the artifacts as
file paths, and pass those file paths from the
reported_step recipe module.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/726165
Original-Revision: a3466a8fce0e648cad8ddae0044bc3069767c656
GitOrigin-RevId: 50418e3a2304e3f44ea4d9ed4004d2aedc33700c
Change-Id: I43211901dfe21e83d2d6de1eb6b8c0f639930c87