[roll] Roll fuchsia [starnix] Serve view backed by framebuffer
This change makes Starnix serve a view backed by a framebuffer for
components that specify the `framebuffer` feature.

The framebuffer is always created and stored on the kernel, but is only
exposed in galaxies with the `framebuffer` feature.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727708
Original-Revision: fea75e1e986a67fbdc0db000af199992d7e61d9e
GitOrigin-RevId: 15b866fdee60fec5333f90a780c72521190998c0
Change-Id: Ieb8397ae84aa7792297895720b62b21f23a8ca50