[roll] Roll fuchsia [starnix] Don't trash mappings in mprotect
Previously mprotect would replace the range given with an expanded
version of the first mapping in the range. This is wrong if there is
more than one thing mapped in the range.

Fixes weird issues in art, which turned out to be a result of calls to
mprotect that covered multiple mappings.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727189
Original-Revision: 183d2bbe028b439cb44616c9fd3caca8b88ffa59
GitOrigin-RevId: 8f989c12a91bc0a0fa759fa32fb7b4a089f036e1
Change-Id: I338750c644ac05a20ca31f7f25f7be215cf06fca