[roll] Roll fuchsia [kernel][mmu] Extra checks and logs for arm64 PTs
Expands the check in `first_used_page_table_entry` to attempt to
validate that the entry found is actually a valid entry, instead of
just not an exact invalid entry. This aims to check for the scenario
where an entry has been corrupted and so it fails a full comparison
with zero, but it doesn't actually have a valid entry.

For the Destroy path an additional method `DebugFindFirstLeafMapping`
is added to attempt to find and display the page table entry, valid or
corrupted, that is holding page tables allocated.

Original-Bug: 79118
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727582
Original-Revision: c8574ef918524d2e47b2120749e87d9349e9fcca
GitOrigin-RevId: 25a864017d422333722e576db6f92bafc45a8787
Change-Id: I57ea8ede13166e6d6857499bdea08df35d635eb8