[roll] Roll fuchsia [netstack3] Move try_update_addr to entry API
This allows doing fun stuff like returning a new Entry to the moved
value on success, which can then be used for further manipulation. This
will be used to update the IP socket of connected UDP sockets when the
socket is bound to or unbound from a device.

Original-Bug: 108663
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726466
Original-Revision: 09c48e24a16f6e4c09c679f6550a63517a387938
GitOrigin-RevId: a3a18c4135ecb9838ca313cf8c92b652be6bbf60
Change-Id: I493260f3a402ab8c2d61238521301b3c5ee937ed