[roll] Roll fuchsia [dfv2] Fix node Remove lifetime
There is a flake where the node is freed halfway through
the Remove function.

The callstack looks like:
- Someone calls Remove on the parent Node.
  - The parent Node calls remove on its children.
    - The last child see's that its parent doesn't have any children
      so it calls Remove back on the parent.
      - The parent's new Remove function frees it.
    - The child is freed.
  - The parent continues the Remove function but it's been freed
    (segfault).

This is a simple fix, all we have to do is keep a shared pointer
alive during Remove to make sure we are never freed in the process
of calling this function.

Original-Fixed: 109777

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727030
Original-Revision: ca3eca051fdd45179e08ce74a6c2535872f69021
GitOrigin-RevId: f0ea140e9eddeab2b245be02048b88ba0632c1c9
Change-Id: I5963c0d33a16b508e93441583cb5a6049f4b32fd