[roll] Roll fuchsia [utest] Set fake UTC backstops to zero.
These tests use ZX_CLOCK_OPT_AUTOSTART, which creates a clock that is a
clone of the monotonic clock. What is less clearly documented is that
the backstop time provided must be in the past for the monotonic clock.
The upshot of this is that these tests currently rely on the fact that
they were run at least 123,456,789ns after the host boots. As it
happens, we spent a not insiginficant amount of boot time allocating an
aspace containing a single 16GiB mapping for EFI. A future CL will
change this, which causes these tests to start flaking (or failing
outright).

This CL reduces the backstop to 0 so that the tests pass regardless of
the time the kernel takes to boot.

Original-Bug: 110001
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726882
Original-Revision: 8b9955b5b73bbcc4ff8f6235922850474b0f4382
GitOrigin-RevId: e75d20204cbb7a871cd2e5f4f2a57bbc52a91a47
Change-Id: If26a4e3b0a172747f3cb150e6631caad8689b59c