[roll] Roll fuchsia [fxfs] Implement syncfs
syncfs(2) is implemented in fdio as a call to sync() on the passed
directory fd.

Fxfs currently does nothing for sync() on a Directory; fix that to
execute a filesystem sync.

For consistency with minfs, we also flush the device, although it's
debatable if we need that.

Original-Bug: 96085
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728323
Original-Revision: cb579256c39d41470dbf8df79620d377f2dcc542
GitOrigin-RevId: 790c517a40f23d5cf93f7f573dccd6c469da33d6
Change-Id: I4a45197e2e1ec4c1abf8027f1b611672109050fd