[roll] Roll fuchsia [rfc] RFC-0138 remove magic number change.
Update RFC-0138 to remove the specification that the change should be
rolled out with a change in the FIDL magic number.

This is done because the design as specified is already backwards
compatible without needing any special magic-number handling logic.

An alternative approach would be to have protocols use the new magic
number to try to provide improved backwards compatibility. However,
this would only help in a very small number of additional cases.

For an updated client communicating with an outdated server, having a magic
number change would break forwards compatibility, because outdated servers
reject all messages unless they have the exact magic number that the
server expects. So for updated clients communicating with outdated
servers, changing the magic number would actually reduce compatibility,
unless we build in special-case support for the new magic number for
builds which don't support unknown interactions yet. This also applies
to server-to-client events.

The only case where changing the magic number improves backwards
compatibility is for an outdated client communicating with an updated
server, where we could use the new magic number to have the server
switch whether it replies with the result union or not in the case of a
flexible method without error syntax. However, the utility of this
extra compatibility check is questionable, as changing a method from
strict to flexible is a backwards incompatible change anyway, so it is
not expected that outdated clients will know about any flexible methods.
Therefore it doesn't seem worth the extra complexity this would cause in
the reply sender code in every binding language.

In the case where an outdated client calls a method which is not known
to an updated server, changing the magic number doesn't affect backwards
compatibility, because all method calls made by outdated clients are
seen as `strict` since the flag used to mark strictness is currently
unused and defaults to 0, and the flag is 0 = strict, 1 = flexible. So
regardless of a magic number change, the server already knows to reject
unknown calls from an outdated client. This is also true in reverse
regarding events sent from server to client.

Therefore I do not think that a magic number change is necessary or
desirable for adding unknown interaction support and am proposing to
remove that from the RFC.

Original-Bug: 88366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722103
Original-Revision: a84201f8137da56f8cd34a97a312bedea440f27e
GitOrigin-RevId: 1a39e9d515ceebcf93d3113fa3123eef57a2df07
Change-Id: Id638077cefb6b4213f657b315c85098759d65264