[roll] Roll fuchsia [build][ctf] Remove deps from verify_deps_in_sdk
This CL changes verify_deps_in_sdk, a template that verifies deps, so
that the verification target it produces does not itself depend on the
deps. To clarify this, renamed the argument from deps to deps_to_verify.

The verification checking only depends on information produced at gn gen
time, so it's unnecessary to introduce edges in the ninja graph. This
was also causing problems for https://fxbug.dev/108354 because the
verification target was getting expanded in the default toolchain, but
the deps did not exist in the default toolchain.

Original-Bug: 108354 (enforce _hlcpp suffixes)
Original-Bug: 87864 (unwanted dependencies in CTF templates)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728085
Original-Revision: fcd986a533695d8eb109955c549e6e36525a65e9
GitOrigin-RevId: 79bbda4d65c60980f3014178e60e9abca62f403e
Change-Id: Ia9880d8ba32dab5b536f641c8e997c97cca5b10b