[roll] Roll fuchsia [netstack3] Update conn IpSock address with device
When SO_BINDTODEVICE is set for a connected UDP socket, attempt to
update the device as well. If the result is unroutable, reject the
attempt to set the device. This is different than Linux behavior because
it makes setting SO_BINDTODEVICE fail in a case that Linux would
allow.

Note that this needs to be documented per fxbug.dev/109358.

Original-Bug: 108663, 109358
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726468
Original-Revision: 4bfe3d4223cc352abef92204a2fca622bf971c43
GitOrigin-RevId: e17fc098601a7532b84550f1629d5f8404ec6836
Change-Id: I1bce2f72eea7a60ebbe8411a0e46e0500a290b4d