[roll] Roll fuchsia [wlan][policy] cancel autoconnect on new connect
Cancels any attempts to auto-connect to a network when there is a new
connect request. In some tests, a connect and disconnect could occur
while network selection is happening for auto-reconnect. In real life
this is highly unlikely to happen because the scans for the auto-connect
and manual connect couldn't happen at the same time. And the
auto-connect wouldn't replace and completed connection because it
will only connect to an idle iface.

Test: added unit test (fx test wlancfg-tests)
Original-Bug: 108037
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724464
Original-Revision: 8a94847729154a487d27a7701fa89dbf7d4b4ebf
GitOrigin-RevId: 571b3cdf695f5a7385a81b2e4ecb90eeee114ab1
Change-Id: Ia3b40dab3b6039cd11920c68fd1635d898ee3174