[roll] Roll fuchsia [dhcpv6] Process Reply to Renew
Process reply message received in response to a renew message, per RFC
8415 section 18.2.10.1
https://datatracker.ietf.org/doc/html/rfc8415#section-18.2.10.1.

Transition to AddressAssigned state on successfully renewing addresses.
Stay in Renewing and rely or retransmission timeout if an IA (regardless
if it is Assigned or ToRequest) is missing in the Reply.

Note: the No Binding case is handled in a follow up CL.

Test: fx test //src/connectivity/network/dhcpv6
Original-Bug: 96319
Original-Fixed: 96321
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/684349
Original-Revision: bd62407194c6b7eb82739bd201640fcf3b6982f3
GitOrigin-RevId: 5b77bc818091b3c9fca86648f90e06891dbf05e1
Change-Id: I26fd73e3c30443d90a34bba7a92e7ac41bcbd74e