[roll] Roll fuchsia [test_manager] Fix events routing in tests
Events routed to tests should be scoped to the test_root. There
shouldn't be both an event carrying the "test_root" name and an event
without it routed to tests.

Fixing this has the nice result of being able to delete the
diagnostics.rs moniker rewriting. However we can't do that until
the fix for the diagnostics CTF test lands and CTF releases catch up.

When the archivist receives a "." moniker it'll interpret it as "<root>"
for selector and presentation purposes.

Run-All-Tests: true
Original-Bug: 81980

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726942
Original-Revision: 9be2f149f6b5540d914ea10f721a0c41e236c414
GitOrigin-RevId: ec2e37a304432d99a3600d040bed423ab3c4eae7
Change-Id: Ie9deeb75e16470a96c2d3a2d5733231b986ddb94