[roll] Roll fuchsia [netstack3] Give `do_send` public visibility
`do_send` was a `pub(crate)` function which depends on `TcpSyncContext`.
It needs to be `pub` because it will be used in a child CL, with change
I59922b0302bf27d77359ea5fc7a4301a3624419d, exporting this function needs
a method in `TcpSocketHandler`, which is done in this commit.

Test: No behavior change

Original-Bug: 48578
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728670
Original-Revision: 8250424553bb01c3b2b50f7c2f82e78864fe0a17
GitOrigin-RevId: ba259e7e5cc0ead3a54fac133eac6b4080d56440
Change-Id: I1740ff7f328a49c85a5b8265661a4edfcaa06e3a