[roll] Roll fuchsia [nxpfmac] Repurpose MoalContext into DeviceContext
With functionality being placed in different components the code base
has ended up in a place where a lot of pointers are being passed around
to access the most commonly used objects such as the event handler and
ioctl adapter.

Repurpose the MoalContext to be a DeviceContext to hold these pointers.
This context can then be passed around instead of a whole bunch of
pointers. Use the context for WlanInterface and Scanner construction
initially.

Test: Compiles, tests run
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720265
Original-Revision: 588589efb4fc2e2810de052004d1fb0553cb82d6
GitOrigin-RevId: 6107badd7d1cd3b13d781c58a6ab6654a8102982
Change-Id: I921b52aa7d0ba4a7a59485ee453139cb64fe9fe0