[roll] Roll fuchsia [starnix] Merge adjacent equivalent mappings in the RangeMap
Adjacent mappings with the same backing store, offset, and protection
should be merged. This is important for the semantics of mremap, which
returns an error if passed a range that contains multiple mappings. This
also affects the contents of /proc/pid/maps.

Fixes allocation failures in AtexitArray::expand_capacity in bionic.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727190
Original-Revision: e8439fb6f102fe60354ae3098d717dd1d9fcc06c
GitOrigin-RevId: 63ad3d352e50ea48aff38611e43278b566f3309e
Change-Id: Ia2bf60e55bf6609b3c71c5444ebb3a5ceaf5d9dd