[roll] Roll fuchsia [fastboot] Add delay before reboot
Some times the reboot is so fast that the devices hasn't been able to
sent the response packet over the transport. The consequence is that
host side fastboot hangs waiting for the response. Add a small delay as
a mitigation to make sure the response packet is sent.

Original-Bug: b/217597389
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728862
Original-Revision: 1025610a58d085ed2032ffe7341500906a56eea3
GitOrigin-RevId: ccb9449277d9914ca0719096f7be8a581174ea61
Change-Id: I5a1fcfe269b7b447dff55255115d75c28f177bfe