[roll] Roll fuchsia [camera] Add ActorBase and associated tests
As part of our efforts to clean up the camera stack, we'd like to
make any asynchrony and/or concurrency in our code very explicit.

An actor is an object which is associated with some means of
scheduling work (in this case an an async::Loop/Executor), and only
code which is scheduled by the actor on that loop/executor is
permitted to read or write the actor's state.

This means that (with some exceptions) an actor's public methods
should create a promise which is scheduled on the actor's executor
rather than run immediately on whichever thread called the method.
These methods may return bridged promises so that other actors/code
can know when the scheduled task has completed.

Two actors may schedule onto executors for the same thread/loop, but
if the paradigm is followed properly the code will remain threadsafe
even if the actors are subsequently moved to executors that run on
different threads.

The base class defined in this CL holds an actor's Executor and
provides several convenience methods so that the child class doesn't
need direct access to its own dispatcher. This is to discourage
passing dispatchers around to random code, and also so that every
scheduled promise is wrapped in a .with_scope(). This means that if
the actor object is destroyed, the work it's doing will be cancelled
and bad memory accesses prevented.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723442
Original-Revision: 110c0a4b70c0bc2e3702deb22317716b97ea1b5a
GitOrigin-RevId: 633e7c2934f250c12f904d5e9475af90efe53cfa
Change-Id: I7488d9ab4ae8cdec9bfeec074e416b5c85c0e469