[roll] Roll fuchsia [assembly] Core product assembly config is now testonly
The core.gni file is not, but the product assembly config
is testonly, which allows it to contain testonly items.

This required forwarding testonly through to a number of
other templates to properly mark them all as testonly
when the product assemblyy is testonly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729323
Original-Revision: 98521642b58a3a2cb98e34d3195475689dcf39e5
GitOrigin-RevId: 2e86e15a7f62e023cca0de45d72488b53de0e20c
Change-Id: Ibfb800050fe50f709ddae6f09f8e4bb7209e9e01