[roll] Roll fuchsia [virtualization] Reduce log spam when shutting down VCPU
When the guest manager is killed (such as in an integration test), the
VMM is stopped and thus each VCPU fails to post the DestroyAndRespond
async task.

This is fine, since the VMM component is going away anyway, but it's
distracting having every single VCPU thread warn us about it.

Tested:
1) Ran debian
2) ffx component stop /core/debian-guest-manager
3) Checked logs

Original-Bug: 104989
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729382
Original-Revision: 0dbb6ef53dd9248db085621517345437c8c5ace0
GitOrigin-RevId: 2b9fd744f9d09253bafa01dcede70673181ecea6
Change-Id: I1fbb71789adedfab562d0d16c63607f77fd1a85e