[roll] Roll fuchsia [netstack/bench] Detect regressions in throughput
Currently the netstack iPerf benchmarks emit a throughput metric without
a specified unit, which means Catapult does not know whether bigger is
better or worse (it's better, in this case). Instead, emit "bits/second"
for the unit and add support in the Catapult converter to understand
"bits/second" as a "bigger-is-better" unit.

Original-Bug: 108749
Test: fx test catapult_converter_unittest
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728960
Original-Revision: 5f14988cbcc45c6191d40351fc46e14d05217840
GitOrigin-RevId: a828c7f61d40c95507a7f63d11675229a6b47870
Change-Id: I830f252a192a1676f17ace3f11e196f94c43a680