[roll] Roll fuchsia [fdf] Create DFv2-specific core bot tests
Create buildbot groups `//bundles/buildbot:core_dfv2` and
`//bundles/buildbot:host-tests_dfv2` which are similar to their
counterparts, `core` and `host-tests`, except they do not contain tests
that fail when DFv2 is enabled and using chromebook.

These groups are to be used by the core.x64-dfv2-release and
core.chromebook-x64-dfv2-release builder bots in a future CL so that the
builders can pass and be monitored by the Fuchsia build gardeners.  This
is in order to prevent regressions in DFv2.

Original-Bug: 102234
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719645
Original-Revision: c4c94ce08cb55104fd066503230161c69209cfda
GitOrigin-RevId: 249a255469ae7fd9a03a59f1ab662d0b0a0b5013
Change-Id: Id4c73f62ee5fcca928ba31ee61194f30de57217a