[roll] Roll fuchsia [ffx] Add back `ffx component run-legacy` plugin
The plugin `run-legacy` is restored exactly as-is, with no
modifications.

This plugin is used by Chromium folks for running v1 components.
It is currently causing the SDK rolls into Chromium to fail
due to Chromium's dependency on `ffx component run-legacy`.

The dependency was not caught at the time because the LSC presubmit for
Chromium was broken when commit 50e4938648c28379e227aa7b17d0776ef7e61873
was merged.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729607
Original-Revision: 0f723099d4cbceaf95fea0b3630cd723985ad66d
GitOrigin-RevId: 03756c34a18950bc646f8c3a1776b8cfd6eba335
Change-Id: Icf8efdfd6ad2b91fa1cbed31393d71aac5d53142