[roll] Roll fuchsia [nxpfmac] Fix flaky dataplane tests
There was a small chance that dataplane tests would flake during
destruction because a raw pointer access. If the pointer was already
deleted the destructor would access freed memory.

With recent changes to the DataPlane destructor this complicated setup
can be simplified. Now we just need to wait for the async remove watcher
thread to complete after deleting the dataplane. Also use a shared
pointer to the device object to make sure it's kept alive for as long as
needed.

Test: Repeated unit test runs with asan enabled
Original-Bug: 110221
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729624
Original-Revision: a62628c43b18ee8905c7db8b2ee9d82af820b763
GitOrigin-RevId: be78dfa9d8c7b061e95c598ff93229d5cc48a2cb
Change-Id: I637c1f7066648e95bae9335e9306eee0357a18c8