[roll] Roll fuchsia [ffx emu] Use `cache=unsafe` for drives to reduce the physical disk writes
The definition and choices of `cache` parameter can be found at
https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/7/html/virtualization_tuning_and_optimization_guide/sect-virtualization_tuning_optimization_guide-blockio-caching
(There must be an official one in qemu.org, but I cannot find it.)
In short, `cache=unsafe` disables the flush command to write-back the
caches in both host and guest to the physical disk. The data may be
lost if the emulator shuts down unexpectedly. But in practice,
there is no reason to preserve data in the qemu instance and
writing-back is a waste of the SSD lifespan.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729186
Original-Revision: 21fa2a7fcd59d2d9ab76265ccde8170594678dac
GitOrigin-RevId: b16542f649884e185027009cf7d8d399a493091e
Change-Id: I516183b8635ebc5a937fd7d01c7f0ebd89026eb3