[roll] Roll cobalt-registry [bt][avrcp] Update metrics for logging player capabilities.
Update metrics so that 0 event code is reserved for Unknown values, so
if an event code for a dimension is not passed to the Cobalt logger, the
value zero can signify meaningless/unknown data. The change is safe
since the metrics is not instrumented right now.

Original-Bug: 105464
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt-registry/+/727942
Privacy-Approval: Cameron Dale <camrdale@google.com>
Original-Revision: 73e686ffea902dd88e2854d20714471bd1e3f99e
GitOrigin-RevId: 1bb149bc64b73b546dbfc3230a8ea6353057316e
Change-Id: I1b4d6a18af87b2ec349204498c8a07d207415189