[roll] Roll fuchsia [build][bazel] Add //build/bazel:generate_fuchsia_sdk_repository
Add a new GN target that ensures that the @fuchsia_sdk repository
is properly generated, and creates a symlink to it with a stable path
if $BUILD_DIR/gen/build/bazel/fuchsia_sdk.

This is necessary due to the fact that the exact repository path is
an implementation detail of Bazel and can vary based on a multiple
factors (i.e. whether BzlMod is enabled, and how repositories are
generated using extensions).

Original-Bug: 106543
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/730783
Original-Revision: 15d9c91202637c831a9e60b2440a301691c9fc3e
GitOrigin-RevId: 74551d5fde770f4e4720ca9aaa8f5fa80aee8838
Change-Id: I245ce7299be3b1b7534d4a26658e41aab87e7472