[roll] Roll fuchsia [driver_manager] Pass banjo symbol from DFv1 to DFv2
Support banjo in the DFv2 alongside implementation.
This is done by passing the symbol from a DFv1 driver
out to DriverManager, and filling it in to DFv2 drivers that
are colocated.

The lifecycle test is updated to call it's parent's banjo
function.

TEST: fx test dfv2_v1_lifecycle
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727344
Original-Revision: 2d0d523447cc460588a9e1b2c244beb67386f717
GitOrigin-RevId: 60b2229f58e7343038ee5b6ef5f8a694f9eebc5f
Change-Id: I8fc5f357d03434306a5a6f9500f826e66f96b45d