[roll] Roll fuchsia [component_manager] Fix stress test total duration
The test duration was set to the string value "60 * 60" in the
test manifest.

The value is expected to be parsed as an integer, not as math
that needs to be evaluated. So it defaulted to running for 6 hours.

Reduced the time to 1800 seconds (30 minutes) because the logs
produced from this test were very large and symbolization would
hang, failing the stress test builder.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/730844
Original-Revision: 1a6d0b9ef292273f116bc052519045deb009f478
GitOrigin-RevId: 12b6aeb8478fc8f525a400a232994697c4deb0c8
Change-Id: Ifd28faf117fc6e2506fdac8e9c49683e274671a3