[roll] Roll fuchsia [zxdump] Read support for zstd streaming decompression
If a file is compressed with zstd, decompress it on the fly.  This
applies to input files and to files embedded in archives.  Note that
reading compressed files is always streaming, so it does not support
random access reading (i.e. zxdump::Process::read_memory).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455254
Original-Revision: 636c5f77130339744957c9fb717be95ea5b6d9b9
GitOrigin-RevId: 11534d6962d3d85324dadf21a1180e84b6480392
Change-Id: I34433528774f739a62a4cf19a1f690bfc448dd08