[roll] Roll fuchsia [fxfs] Unlock volumes read-only for fsck
Per-volume fsck was not re-locking the volume after checking it, which
is problematic for two reasons: (a) it's a possible security issue, and
(b), it prevents the volue from being mounted (which prevented this from
actually being used in a real environment anyways).

However, we couldn't just re-lock the volume, since fsck holds write
locks on the entire filesystem and re-locking triggers flushes.

Fix this by having fsck unlock the volume as read-only, which prevents
all flushes from occurring on the volume.  The encrypted mutations in
the journal are buffered in the lock state of the store while it's
unlocked read-only.  fxbug.dev/110284 describes a more graceful way we
can deal with this later.

Original-Bug: 106845
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729483
Original-Revision: c071d098ab06693efb1d5ccae5a60a984af6b53a
GitOrigin-RevId: 26445fcb63ceccf8eedcb4457f43f3246740cb7a
Change-Id: Ib694a1bc3e03c4bd5f8d0f8af5560e8edca9a4f7