[roll] Roll fuchsia [storage][perf] Add new benchmarks
These new benchmarks are intended to replace the current set of
benchmarks. The current benchmarks are not being removed in this change
to allow for some overlap between the two sets and ensure that the new
benchmarks are working as expected.

Original-Bug: 94248
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698443
Original-Revision: ead8c8bf4d2decfc6b51bdd19921bfcc9ca8255b
GitOrigin-RevId: 0c4900de9bb119f81ed1f415a3faa630fef497f6
Change-Id: I9d23bcddba3c2a2090cf1eae18ab700adb91d999