[roll] Roll fuchsia [ui][input-pipeline] gestures: rename details
Rename `MismatchData` and related types. This is in
preparation for using the types to carry explanations
of why a gesture ended.

Although we could call the end of a gesture a "mismatch"
as well, doing so would likely be confusing: readers
could reasonably assume that a mismatch is related to
the `Matching` state.

Original-Bug: 107948
Test: fx test input_pipeline_lib_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728823
Original-Revision: ebf896a9d1d658edcd85e7ec2a58d0472cb60c2e
GitOrigin-RevId: 3e09905183366d0911af6628ea71fd87d7b522bf
Change-Id: I0eb7e38a34e9da198220007de4ad2988ed54e6c2