[roll] Roll fuchsia [storage][fuchsia-fatfs] Update fuzzer corpus location
As part of the v1->v2 migration, it was discovered that the
fuchsia-fatfs-fuzzer corpus emitted to the manifest didn't match the
location within the package. Previously, this resulted in a fairly quiet
warning; in v2 it is upgraded to a CQ failure.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/731324
Original-Revision: cada06ef30a7e81522f3fecbaebf5941a923d5ef
GitOrigin-RevId: 369a1dee2f3b59bd875d3b8fcea40864323b9430
Change-Id: I01a7ab6b859acd712c1d0c70448a2e27a8af3ab2