[roll] Roll fuchsia [nxpfmac] Don't complete destroy interface early
The wlanphy DestroyIface call would call DdkAsyncRemove on an interface
and then complete the request. However the interface is not completely
torn down until DdkRelease is called on it and the interface destructor
completes. This caused problems when a fast sequence of DestroyIface and
CreateIface occurred. The interface was not completely destroyed when
the create call arrived. Even worse is that the CreateIface call would
succeed but silently fail to create the network device port, thus
preventing any data from being received or transmitted on the interface
will still appearing to be successfully created.

Fix this by ensuring that DestroyIface does not complete the request
until the interface is fulle destroyed.

Test: ACTS test_start_stop_client_connections, manual tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729184
Original-Revision: 08bf285bd5160b19992992971388330d665fbfbd
GitOrigin-RevId: 234bded62bc0ba08d8e44eba4952e06dbcd5f60b
Change-Id: I2cae989493a217e5431618a124795ead6d8bd569