[roll] Roll fuchsia [nxpfmac] Only complete DdkInit once FW is ready
DdkInit would synchronously reply to the init call even if firmware
initialization was in a pending state. In situations where the wlanstack
is already up and running when this happens this meant that the
wlanstack would start making calls into the driver that would attempt to
use mlan before it was ready, leading to crashes as mlan tried to
dereference null pointers.

Fix this by detecting when firmware initialization is still pending and
deferring the transaction reply until the initialization completes.

Test: Induce a crash in the driver that causes a rebind
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729185
Original-Revision: 5e43c33dabe0e95ca3763cb7bb8f35859128750b
GitOrigin-RevId: 2585c0f2e7b5b59171f85d548ca79b0c3033c567
Change-Id: Ibd954b19e02c28e93a09441e7b4392a2239979c6