[roll] Roll fuchsia [pbms] Structured UI for product-bundle, take 2
This is a re-working of fxrev.dev/665178. In that CL, the structured UI
module was introduced to the PBMS library for FFX, which provided much
cleaner, more user-friendly output for calls to `ffx product-bundle`.

This CL takes the original change and merges it with the current
code-base, including some minor changes to account for PBv2 code, etc,
and also addressing the outstanding comments for that previous CL.

Original-Bug: 96950
Original-Bug: 109933

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729490
Original-Revision: 8d3c852f4842ba680b09891b0aab79d6c28b579b
GitOrigin-RevId: 1c3caf98be70dd50c89417443ad72dd4e2e4bd95
Change-Id: I177bb8f057151b26830a7db8ad8a109cf8785ccd