[roll] Roll infra/recipes [testing] Only consider the last result to determine failure/flake.
https://go/bbid/8801936466011918577 is an example where the build
should've failed because one of the multiplied affected tests failed.
However, we considered it a flake because it had some passes and we only
considered shards starting with "multiplied:" as multiplier shards.
This change looks at the last run of a test and determines if it's a
failure or flake depending on its result.
Test: https://chromium-swarm.appspot.com/task?id=5d98e037dc8a2110

Original-Bug: 76711
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/731802
Original-Revision: 2d66c862edf92c152a0fd06686e2615ba7aa494b
GitOrigin-RevId: c08e644f529852de3e1f66cdf968e6e0ab90f21c
Change-Id: I8e1c444559481661415e3f7bf12163962202e630