[roll] Roll fuchsia [power][test] Fix flake in driver_manager mock
The driver_manager mock will panic if it gets an error sending a
response to the client (Power Manager) for a SetTerminationSystemState
request. We shouldn't panic in this case because it simply means the
Power Manger has already closed the channel (exited), which happens as
the RealmInstance is destroyed.

Original-Bug: 110505
Multiply: //src/power/power-manager/integration
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/731768
Original-Revision: 78b07cf57c00935c728e94211096ad2fe4725b84
GitOrigin-RevId: 1edfc09999caf594decd68234b044bfc663ec4c5
Change-Id: I96d955709291751e6d759e51b4a9b196c7c8d3a0