[roll] Roll fuchsia [power][test] Fix flake in driver_manager mock
The driver_manager mock will panic if it gets an error sending a
response to the client (Power Manager) for a SetTerminationSystemState
request. We shouldn't panic in this case because it simply means the
Power Manger has already closed the channel (exited), which happens as
the RealmInstance is destroyed.

Original-Bug: 110505
Multiply: //src/power/power-manager/integration
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/731768
Original-Revision: 78b07cf57c00935c728e94211096ad2fe4725b84
GitOrigin-RevId: 1edfc09999caf594decd68234b044bfc663ec4c5
Change-Id: I6e5b23b868e4a3df91b1ac3eed12f42762c6141f