[roll] Roll fuchsia [overnet][hoist] Work around short SUN_LEN issues
This makes it so that when we try to use a long path to a socket, we
temporarily create a short path 'landing link' in TEMP to bind or
connect to it through. For the most part, we still talk only about the
real socket path and only use it for the connecting/binding calls, so
that it doesn't interfere much with other things.

Note that it actually creates a symlink to the directory the socket is
in, not directly to the socket. This makes sure that we behave properly
in terms of creating the socket file and probably makes it more likely
this will eventually work correctly on windows (though that will likely
come with other challenges, and the likelihood is still low).

While I was in here I cleaned up some of the dependency logic and
simplified the gn template and use statements a bit.

Original-Fixed: 109815
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/731737
Original-Revision: 4b21e40703da815e33b4e83b3c2124db31511123
GitOrigin-RevId: dfa57cb4e3ece3215f246890f79c9ac13610a52e
Change-Id: Ifc70cff764a4bfbcbeddbf264fe8a41fe504bc76