[roll] Roll fuchsia [wlan][hw-sim] check security in connection status
Previously, when tests checked the status of connection through WLAN
Policy client state updates, only SSID was checked. After this check,
the security type is checked as well.

In particular, this would help in a flake from fxbug.dev/108037 where an
auto connect could result in a connection succeeding as a network is
removed. In that case the 'connected' update would be counted toward a
following test, even if the security type used in the next test is
different. Checking the security type would make that test fail sooner
so the failure would be easier to triage.

Original-Bug: none
Test: ran hw-sim tests
      (fx test//src/connectivity/wlan/testing/hw-sim/test)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728628
Original-Revision: c5cff5bcfbc564ebf91ef1b6075823c56dd28c81
GitOrigin-RevId: 8d48d906d3191a4060313e26c3946aef4bf1349b
Change-Id: I917dac7d97bc48fcbae0b3a9f9504923e3eba5ff