[roll] Roll fuchsia [activity] Remove the implementation of fuchsia.ui.activity.Tracker in the legacy activity service.
The protocol API is not removed as part of this change since that
requires going through the LSC process, and will be done so separately.

This change does not remove the implementation of other protocols served
by the legacy activity service, namely fuchsia.ui.activity.Provider and
fuchsia.ui.activity.control.Control. This is because existing users rely
on this functionality, and the service cannot be removed entirely until
they are migrated off of them.

Original-Bug: 103843

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714982
Original-Revision: d79352c03fa0e5b9b2579d146e4295f7373cef21
GitOrigin-RevId: 14ee16994cc99c76aed95783b7cd10ec98e1332d
Change-Id: I59e7084a4d5d980ac9bf6c78dee40cf9aa8bac0d