[roll] Roll fuchsia [power][test] Allow disabling temperature filter
Adds support to disable temperature filtering in integration tests. The
switch is added to Power Manager's structured config, and is always
false by default.

It's useful to disable temperature filtering in tests because the filter
is dependent on time which is not predictable in a testing environment.

Original-Bug: 104266
Test: fx test //src/power/power-manager
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727970
Original-Revision: 534fd905ad567bab35c8e5634391d7fc6a7989f3
GitOrigin-RevId: 4047829c2e0b9f15e33111416ec38af504d5ae4c
Change-Id: I585f710ec94b3e77981424a2740aa5e140fdee37