[roll] Roll fuchsia [audio] Split Thread into `fidl/` and `mix/` parts
Graph nodes are split into a class that lives in `fidl/` (Node) and a
class that lives in `mix/` (PipelineStage). The Node objects are updated
eagerly from the FIDL thread while the PipelineStage objects are updated
asynchronously.

This CL adds a similar structure to Thread. Now we have GraphThread (in
`fidl/`) and PipelineThread (in `mix/`).

Splitting threads in this way makes some graph operations simpler. For
example, Graph::DeleteThread needs to know if there are any consumers
still using that thread. It can't ask the PipelineThread because that
object is updated asynchronously. The simplest approach is to manage
that accounting in both `fidl/` and `mix/`.

The main changes are in the following files (other changes are mostly
mechanical):

mix/pipeline*thread*
fidl/graph*thread*
fidl/testing/fake_graph*

Original-Bug: 87651
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/731302
Original-Revision: 8512ef541d1c2f3b5902103daf757338cb963fe4
GitOrigin-RevId: a7f501386f2b38d27e58c88bda8eac9774ff52f2
Change-Id: Id57d0a206f682aa02e09c78ee19aebc021e10463