[roll] Roll fuchsia [dfv2] Put composites at the top of /dev/
Make DFv2 composites appear in /dev/ at the same place they
would appear in DFv1.

In order to do this DFv2 has to pass the name of the node
that the driver is bound to. This is different than the
topological path created by the parent drivers, it has to
come from the framework.

Original-Bug: 106517
TEST: composite-driver-v1-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711247
Original-Revision: 43febb710b6455170e3cfa86f47f0d817a2500b8
GitOrigin-RevId: b98113a231c21239827d89b5826d90a346fc0541
Change-Id: If2b213ee6e013d6eb0204e000f7cadad32cee4a5