[roll] Roll fuchsia [ui] include fuchsia.media.sounds.Player in the minimal ui realm
Scene Manager (namely, Input Pipeline) will use this protocol to play a sound from the FactoryResetHandler. I made this optional so that it can be available in the minimal realm (where it is needed), and omitted from the regular ui realm because it hasn't been indicated as needed before.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/732296
Original-Revision: 7f2f19846b70970e578a97614380ee612109214f
GitOrigin-RevId: cc434488e7494190e8132ef0fa84e0883f093e9b
Change-Id: Id9d2000e2d1e06cb3f2964d626f14b1fe1c3fa13