[roll] Roll fuchsia [build] Fix visibility usage in structured config values.
When restricting visibility of a config value file,
GN complains if the target itself isn't added to the
visibility list. This is caused by the Fuchsia target
being depended upon by the group that's used by non-Fuchsia
toolchains, so we need to list them too. However we don't
have variables defined for all the toolchains that might be
involved, so we add a blanket visibility for the target name.

Original-Fixed: 110496
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/732542
Original-Revision: d2bfbb5f0a438c628d150a4824debb608c6efede
GitOrigin-RevId: 7afe5e04f6336c56651bab2ca8170e06bcc6cc77
Change-Id: I4375e0b9ca57b54b5a3191204c85e96333eaa493