[roll] Roll fuchsia [testsharder] Skip unaffected if affected tests found even if not run.
The build might detect affected tests that don't have an environment
attached and will thus not be run as part of that builder. This will
cause AffectedModifiers() to return 0 modifiers, but the original
affectedTests file did have tests that were detected to be affected. So
we should check that file to determine whether we need to run unaffected
hermetic tests instead of checking for an empty set of affected
modifiers.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/733133
Original-Revision: 3e247cb9d3fb771c18a065a0632fd3b44f955de6
GitOrigin-RevId: 163903281e728e533e1504330d3f6935dfec12c0
Change-Id: I9d07b2e8bbaf02c49bf479ad7d44dc43d701219c