[roll] Roll fuchsia [iwlwifi] Remove banjo dependencies
The iwlwifi driver uses some of the banjo generated definitions and
since banjo will not be available oot, we are copying over these definitions
locally to allow the driver to compile once it is oot.

Note, since these definitions are no longer coupled to their respective
.fidl files, the driver needs to ensure that it always translates values
between the fidl interface and the local ones, and not copy them as is.
Doing so could lead to bugs when they get out of sync.

Original-Bug: 99490
Test: Existing tests cover this change

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/676148
Original-Revision: 8b7fe178bf5097cb0939967187b0f3241742fde6
GitOrigin-RevId: 027292414606a99f17a5f5d264629f795138a959
Change-Id: I7bd9ef0de9ebbea4e1209b0379686c7e20dbe180