[roll] Roll fuchsia [audio] CreateEdge and DeleteEdge move subtrees across threads
When created an edge A -> B, we currently move A to B's thread.
Similarly, when deleting edge A -> B, we currently move A to the
detached thread.

However, we actually need to move an entire subtree of nodes above A.
This is implemented with a new function in reachability.h,
MoveNodeToThread, which operates on a new property, Node::is_consumer,
as described in the comments in reachability.h.

Original-Bug: 87651
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/732567
Original-Revision: 0d2531d33c141c23bb65c5a161942f92dc2eb6f7
GitOrigin-RevId: cada7422e560b2b5e1364006f33040664e4782d6
Change-Id: Iaa046c9adf51360331d11ae1ba48938777aa087a