[roll] Roll experiences [gazelle] Add `wm-tests` to CQ.
I had failed to restrict that test to devices with graphics support. I
got a protip that `environments = vulkan_envs` is the way to do that,
so I've added that to `wm-tests` and to the AppKit tests as well, to
make it clearer what the `AEMU` reference is for.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/731782
Original-Revision: 1f740b79e9d5dd00cccc4e1b3d3db36579447e8e
GitOrigin-RevId: 460ac57713a2c44b301f2f1d8a2438914bd06873
Change-Id: Iff8d9f6bd3b0cbce15419f345a814f1044ab273d