[roll] Roll fuchsia [netstack3] Implement RX device queue for loopback
...so that loopback packets are never handled synchronously with any
send path. That is, this change decouples the TX and RX paths for
loopback, resulting in any call into core only ever sending a packet
or receiving a packet, but never doing both. This allows (in later
changes) for packets to be sent while holding a socket lock without
a chance of a deadlock when the sent packet is expected to be
delivered to the sending socket itself.

Original-Bug: 48578

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/733043
Original-Revision: d8699a735c65efe90753d07b48b167d701162041
GitOrigin-RevId: 7089913deac82363c1c824bc2182699db0b88e7b
Change-Id: Idede21d81addf84055350500fdf6782ef7f2cc04