[roll] Roll fuchsia [intel-i915-tgl] Only write transcoder disable config when transcoder is enabled.
On Tiger Lake, if the driver sets `enabled_target` to false and writes
the TRANS_CONF register when the transcoder is disabled, the whole
display hardware will crash and cause the system to reboot.
Instead we must only write the register when the register is currently
enabled (`enabled` bit is 1).

Original-Bug: 109368
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734044
Original-Revision: 5326f7ae7d633d637fceb78805b8bc3e2a41b34e
GitOrigin-RevId: b0d64e623f857f1ce5e59ec69a0ff66d1d342b7e
Change-Id: Ie33ca6ad4a01c293f8707dd6772debfacfb03625