[roll] Roll fuchsia [driver_manager] Replace Device::IteratorFactory
Clean up the code to get the Device's children.
Instead of using a hand-rolled iterator and iterator-factory
class, we can get the children ahead of time and stick them
in an STL container so we don't have to write the iterator code.

This should have identical behavior to the orginal in terms of
which children are present.

The main benefit is cleaning up the logic and the header file.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/733624
Original-Revision: c4ce986a8f609d210564d563e344f09d41bbb7e4
GitOrigin-RevId: e4dcbb5fbf6b2cb662cdf4b6cb59db5dbe675437
Change-Id: Ibd6d24a2f42bf2d85cd50a07cf8267cfd27ff13d