[roll] Roll fuchsia [testsharder] Don't set StopRepeatingAfterSecs if user specified runs.
Before, we were setting it to the target duration, but that would cause
the shard to exit early before running all specified runs and there
would be no signal that it ran fewer tests than specified. See the
multiplied shard for go/bbid/8801575033527660129 which only ran 49 runs
when the user specified 300 in the commit message. This could be
confusing, so we should just try to run all runs, and if it ends up
taking too long, the shard will just time out which will at least
provide signal that we couldn't run it for the full count (so the user
will need to specify a lower count).

Multiply: web-semantics-test: 60
Original-Bug: 76711
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/733845
Original-Revision: 8dd4ea9deaf2e76f56e87376e4211645505a363a
GitOrigin-RevId: a67b24df064fbbec0fd25a48e41e25dfe1564660
Change-Id: Ia1e9e17093fe9a2f5000db4d2f98d736758a6531