[roll] Roll fuchsia [bt][core] fix std::vector usage in profile_server's DataElementToFidl
A bug was introduced in fxrev.dev/642344 where the incorrect std::vector
constructor is used to copy a string. The code created an std::vector of
the correct size but filled it with the first byte of the source data.
Instead, we need to copy all bytes in sequence over to the new
container.

Original-Fixed: 110224
Test: fx test bt-host-fidl-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729210
Original-Revision: b81038f8a3d4070b0b962947fd577a60ac212473
GitOrigin-RevId: ce2de30ec2e9ed42e1c0b8f24b21d1ad24ff0b55
Change-Id: I65954d151e9b2a6ceb58bbd13976d2f3863b1085