[roll] Roll fuchsia [blackout] use dev-topological for dev capability
In fxrev.dev/722416, the dev capability got renamed to dev-topological.
This was changed for the integration tests, but not the actual tests,
which means the actual tests are attempting to route a capability that
doesn't exist.

To try to prevent a change like this from not being propagated in the
future, organize the repeated set of cml fields into a blackout shard
that is included by the tests instead of repeating ourselves.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/732602
Original-Revision: 8fdcf88555cf27b4dada16c96da2d37abac7e907
GitOrigin-RevId: caf51d75640f228bdb8b8a9c3fd3db0889d035b6
Change-Id: I37d26848341b816145377e2bfdf468e95888c2ed