[roll] Roll fuchsia [netstack3][tcp] Fix a panic in Assembler::insert
The outstanding Vec is non-empty only when start != end. Normally this
won't happen, but when the read side of the socket is shut down but
there are still incoming segments, we will try to insert a range with
start == end, causing OOB access.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734346
Original-Revision: 23ec5376491ebe1212aa60faf9bfa8b75a541255
GitOrigin-RevId: f29d94a8ef1f43ac2876b7419bae3dd64595eac4
Change-Id: I94bd30eebf3e0ecc5da35973be25a069270cf865