[roll] Roll fuchsia [wlantap] Remove ClearAssoc() MAC address NULL check
The MAC address is no longer verified in the mocked ClearAssoc()
function. In the implementation prior to the DFv2 migration in
fxrev.dev/660030, only the array pointer was checked to be non-null,
not the content of the MAC address itself.

Test: hw-sim test suite
Original-Bug: 111067
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734342
Original-Revision: d4cf7b6648b152f025452617117dfdd4af471faa
GitOrigin-RevId: 3f498bd35aa72f368d2de2bd47f34c589bba086e
Change-Id: I4c61b8d77f1f0e8a1a09e4f449ae4b7dbf39b382