[roll] Roll fuchsia [emu] Remove double-listing of PBMs on start
fxrev.dev/718948 introduced achange to product-bundle loading in the
emulator launcher that caused the list of available PBMs to be printed
twice, when a complete URL wasn't provided on the CLI. This is because
the product URL was needed in multiple locations, but the name was being
passed instead (which triggers resolution).

This CL adds a parameter to the select_product_bundle function,
should_print, which suppresses the output when set to false; this is true
in all cases except the known second resolution for `ffx emu start`.
This is considered an acceptable workaround, since the product-bundle
tool is deprecated pending PBv2 completion.

Original-Bug: 110265
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/731736
Original-Revision: d3acd572110a8f9ea5db704d8aa24762ffc019d3
GitOrigin-RevId: cbbc6f2a907a67a2089e63e27e5bd8921258aff4
Change-Id: I711d5898e6b65bb37f5190358d238454d3950d5b