[roll] Roll fuchsia [devfs] Remove Devnode::is_invisible
Prior to this change, nodes were added to the topology on creation, but
kept "invisible" through internal attributes that were inspected on each
interaction. Following this change, nodes are held in an unpublished
state until `Devfs::advertise`, at which point they are added to the
topology and become proper, visible nodes. This means that
`Devfs::publish` is now a misnomer, so it has been renamed.

Original-Bug: 38246
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/732532
Original-Revision: 8cea3c990fe21997adf0a537969952b533063c41
GitOrigin-RevId: 9e33f8f2f22062a2113f452f7b16fb4ca21593ca
Change-Id: I1a0e772cfa8ca054cc366564f29905f637901744