[roll] Roll fuchsia [dfv2] Use node's real name for composite names
Node::name() returns the node's name with the ":" and "/"
replaced so it can be used as a component moniker.
Store the real name in the node so that we can use it
for composite devices correctly, so they show up with the
exact same name they were given in DFv1.

TEST: device-enumeration-test in next CL

Original-Bug: 106517

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/732875
Original-Revision: fd5913e050302ded305e9bad98d8f5a56f4021ed
GitOrigin-RevId: 50fdbd20a5ca3066cf8df06f2962a4397a208aef
Change-Id: I9a8230fbf7b71c75f71af53151f12f31b1d8f4d2