[roll] Roll fuchsia [vm][utest] Drop non-essential WaitForBlocked() from pager-writeback
Most of the tests in core-pager-writeback do not strictly need a call to
WaitForBlocked(), which can be flaky when the test is run as a component
and the test thread is blocked on a pager outside of the test. Remove
such instances of WaitForBlocked().

These tests would wait on a pager request immediately after the
WaitForBlocked() anyway, so the WaitForBlocked() was not strictly
required. The WaitForBlocked() check was not providing any meaningful
information in a component test.

Original-Bug: 109851
Multiply: fuchsia-pkg://fuchsia.com/core-pager-writeback-test-package#meta/core-pager-writeback-test-package.cm
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/733924
Original-Revision: 308807d8c8bb8604e03957edccc753c6bdfa0b43
GitOrigin-RevId: 8db322c194938a2d68b5cbaad1ad43fe091285a4
Change-Id: I8ae10942148c04071f4fa9beb5da6f37a3d7acea