[roll] Roll fuchsia [vm][utest] Skip some pager-writeback component tests
Some tests in pager-writeback.cc were relying on WaitForBlocked() which
will only be correct when there are no external pagers. So these tests
cannot be run as a component. They cannot be run as a bootfs standalone
test either (which is also true for the rest of the tests in
pager-writeback.cc), so these get run only in unified mode.

Original-Bug: 109851
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/733925
Original-Revision: 8e6617e0581bd2bc3bc9b953480711c6d1433194
GitOrigin-RevId: 07fc9f6aaae2ea683580249c969a717238a205a1
Change-Id: I4ccd892f0decf9ff6495620f1fef7b4acc4a7778