[roll] Roll fuchsia [zxio] LRU eviction for dgram socket cache
Prevent unbounded growth in the client-side send cache by keeping a list
of the entries in most- to least-recently-used order. When an entry in
the cache is inserted or accessed, move it the front; when the cache
reaches its size limit, discard the least-recently-used entry.

Note that this CL also increases the size of the storage backing a
`zxio_t` in order to accommodate the `std::list` added to the route
cache in the client.

Original-Fixed: 97260
Test: fx test zxio-unittest-package
      fx test netstack_with_fast_udp_dgramsocket_c_api_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729945
Original-Revision: 915a292975575e186ee6ec5ca1a8a09050d48430
GitOrigin-RevId: 0cd544dfafb05e707e574a25a796b82ad5eb5409
Change-Id: Ia7dc39fd8255d3f73acf66562785269a1ef99c1c