[roll] Roll fuchsia [intel-i915-tgl] Add transcoder chicken reset step.
This CL implements step 2.d. in the section "Sequences for DisplayPort"
> "Disable Sequence" on pages 147-148 of IHD-OS-TGL-Vol 12-1.22-Rev2.0.

This CL does not fix the system hang in the attached bug. However, it
does bring our driver logic closer to the official documentation, so it
is a step in the right direction.

Original-Bug: 111129
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734490
Original-Revision: 9be79679bc507db66acd048a9e7226924d29ec21
GitOrigin-RevId: 482b3037630b85a8234c66260e5edc87e6d584ad
Change-Id: I91a8628add5b0e649ca9642fd73438e2e44aff58