[roll] Roll fuchsia [intel-i915-tgl] Add transcoder VRR reset step.
This CL implements step 2.h. in the section "Sequences for DisplayPort"
> "Disable Sequence" on pages 147-148 of IHD-OS-TGL-Vol 12-1.22-Rev2.0.

This CL does not fix the system hang in the attached bug. However, it
does bring our driver logic closer to the official documentation, so it
is a step in the right direction.

Original-Bug: 111129
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734782
Original-Revision: 654428d3ff2705c245f29a563377af8b59d1095e
GitOrigin-RevId: d61e65a18be10fd509e2b67c8f263b567ec72327
Change-Id: I17d49ed19ee2a8d27f3e3d31b2d19dc0dfadc0aa