[roll] Roll fuchsia [dfv2] Include device-enumeration-test
Include the device-enumeration-test on DFv2 devices.
Update the test so it checks for the DFv2 composites.

This CL outright removes the check for:
pci-00:00.0-fidl

This is a composite that does not have a driver bound to it.
Composites without drivers don't show up in DFv2, and because
there's no driver bound to it, it's likely not worth checking
for in DFv1 either.

Original-Bug: 106517

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/731885
Original-Revision: 9838eeb92b2994567bcdcc7a05aac1688bce2979
GitOrigin-RevId: 3e48dbb426fc59b37fe9502894ca4d32f7a7689a
Change-Id: I73cb049dc0a58db85f13bab007ef64f1b1080bbf