[roll] Roll third_party/pigweed pw_transfer: Support protocol v2 in Python client
This updates the Python transfer client to work with either the legacy
or version 2 transfer protocols.

To support this, a wrapper class around the generated protobuf Chunk
class is implemented, supporting serializing and deserialzing specific
sets of fields based on the protocol version in use and type of chunk.

To set the protocol version, a default can be provided when constructing
a transfer manager. Alternatively, version can be set on a per-transfer
basis when starting a new transfer. As with the C++ client, the default
version remains LEGACY until all transfer implementations are ready to
switch over.

Original-Bug: 243679452
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/111330
Original-Revision: 0ae91ed7077a03d45edbd0ba86f27b3b4da773ba
GitOrigin-RevId: 65c685454d0ed0d524c7d9862e5306e0e26f3cba
Change-Id: I4476cef9eaaff21b839b125a4eb6542ba3515e42