[roll] Roll fuchsia [wlan][hw-sim] Remove unused emergency capability routing
The `emergency_location` component is not used in any `hw-sim` tests
and no longer needs to be routed. As I understand it, the existence of
`emergency_location` in these manifests was based on a previous
requirement of CFv2 to include all component dependencies even if
they're unused.

Original-Fixed: 78327
Test: fx wlan-hw-sim-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735284
Original-Revision: e68a6ce8c2df72ffcacc169d593cc8d69ed47a64
GitOrigin-RevId: 59033bd105ba7a7c33a63849ace98a4aeac12f54
Change-Id: I26595546bad2191eecc8f892dc430e1255ed74b9