[roll] Roll fuchsia [assembly][boards] Add board-provided drivers to product assembly config
This is an interim solution that allows board-provided drivers to be
added to assembly via an internal detail of the
product_assembly_configuration() template.

This is meant to provide a better mechanism that fxr/729083 patch#9
is currently using.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/733962
Original-Revision: 9d48cae735a935f5f4ac74b3608f2266b1fd0da6
GitOrigin-RevId: 4bf175756120ebbe13c2bf76d06ad045a2eff86a
Change-Id: I898bb8f9c2bbe86f2c629bdeff03d504c132ab52