[roll] Roll fuchsia [direct-mode] Optionally pass guest handle to process
This allows us to optionally pass the guest handle to a process launched
within direct-mode. This will be useful in an upcoming change, where we
can use the guest handle to control whether the process is allowed to
act like a kernel and create its own address spaces to manage.

This option is controlled through an environment variable, in order to
minimally interfere with the process being launched.

Original-Bug: 95763
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734826
Original-Revision: 7b52ca98676fbd08d2a6e5a8845c112e0fb9d9a6
GitOrigin-RevId: 5628350e7d0e693d276b0bc3a224c4d6c3386135
Change-Id: Ife9b33be3f0b3cc8ed73a48f17948c7451478403