[roll] Roll fuchsia [bt][intel] Update firmware reset after secure fw load
When loading firmware, we send a vendor reset with a boot address in the
parameters for it at the end.  This boot address was previously
hard-coded but is not static for all intel controllers, so instead read
the boot address from the firmware as we are sending it, and send the
correct address on reset.

Update some logging while we are here (newlines are not required, add
more info to some)

Original-Bug: 108011
Test: boot atlas and observe info in logs from firmware
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/725383
Original-Revision: e1bb1fba4c4f56bbe837e2d857647cd00a016859
GitOrigin-RevId: 961f2b4f3506fa0d545d8108d904f332c573579e
Change-Id: Ifcab93df901a4c1c5d473b33fa2c5ffa6c5df724