[roll] Roll fuchsia [usb] Refactor EndpointList
Copying descriptors is useful for avoiding accidental nullptr derefences
but it masks invalid states in a way that is difficult to debug. In
particular, it's difficult to distinguish a valid endpoint iterator from
an invalid one (such as is created by |end()| or in test data.)

The specific problem this caused was that some existing test data
contained an endpoint descriptor with an address set to zero. Previously
the equality check relied on the USB specification mandating that
endpoint addresses are unique and not zero, and so the test data was
equivalent to the iterator created by |end()| and would not iterate.
This was nonintuitive to debug. It's not the job of the iterator to
validate the descriptors, so we should at least ensure that we can
iterate through them regardless of validity.

Also modernize the endpoint iterator struct with a few convenience
features like using std::optional for the superspeed companion.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/733363
Original-Revision: 5376b74f9e892ff4b1dfcca2b0e538b49375966e
GitOrigin-RevId: 743adebf6d8b157ad8b97b9fe9c8fbe90a9c0c26
Change-Id: I72429d7b208be500cc02859e55d1d485a5c75f03