[roll] Roll fuchsia [virtualization] Add integration test for restarting a guest
This test validates that a guest can be cleanly restarted (including
bringing up all of the devices, etc.), without restarting the full
component.

Since this test is polling guest manager for when the VMM has stopped,
this needs to fix a minor race where the VMM component only reports that
it's stopped after it destroys the VMM object.

Note that this currently skips termina as termina requires that the
guest manager is torn down since it requires /dev, and drops /dev when
starting the VMM for security.

Original-Bug: 104989
Multiply: virtualization_shutdown_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728225
Original-Revision: b895369c5d85375bbcf4c339dfbe88e19ce64696
GitOrigin-RevId: 2c1730493e02afaa95e9e4983cc3293a57b70d89
Change-Id: I82fc17b377a25738c5689b49b86e7d713ab5ff0b