[roll] Roll fuchsia [pbm] Always default to the local bundle in-tree
This CL modifies the product-bundle selection code (called when
disambiguating multiple available product bundles) to always default to
the bundle generated by the in-tree build when it's available. If no
local build is available (which shouldn't be common), it will fall back
on the previous behavior of selecting the first in a reverse-sorted
list.

Original-Bug: 98978
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734987
Original-Revision: 9a297df693ea62b8aa8aa1549c7e1fa104fde0a2
GitOrigin-RevId: 17061886fe5db28adcba8b42d3d1f7d50be708a6
Change-Id: I4aee81273c37506f2cae7982afd96cebfe494ddd