[roll] Roll fuchsia [text-input] An attempt to resolve a test flake
This test was flaking when many repeated taps on the browser screen did not result in a focus change
on the input field.

chaopeng@ and I think that this might happen as a result of (1) the browser window gaining focus before JS is fully loaded (which is corroborated by logs) and (2) the 'autofocus' attribute on the text
field, which means the field will be focused when the page loads. (we couldn't corroborate this with jsfiddle for example, as 'autofocus' seems to behave differently there)

So, if (1) happens, then (2) may not result in a focus change notification, because JS didn't install the focus handler yet when the focus was changed,. In addition, autofocus would have already placed the focus on the text field by the time JS is operational. This would lead to the code never getting notified
of focus gain.

If this hypothesis iscorrect, then the removal of the "autofocus" attribute might resolve the
issue, modulo any misunderstandings of the underlying mechanism on my behalf.

Original-Bug: 111219
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735306
Original-Revision: b7873aadf9de655fa53229739394f09d5b6b5885
GitOrigin-RevId: 5a4119d5077e13f60fe8ee95fd1c4502f3f0082a
Change-Id: Ieba2e158eea00c12107321517e93ebd94c55c14f