[roll] Roll fuchsia [ffx][flash][test] Make test harness more robust
Removes the public "variables" member of FakeServiceCommands, and
replaces it with `HashMap<String, (String, u32)>`. Also adds functions
to set the variable and it's associated value, as well as retrieve the
number of times a particular variable has been called.

This removes the need for developers to care about managing the order by
which the variables are retrieved from the device in their test, and
also allows them to make assertions on the number of times a variable
has been retrieved by the calling code.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734986
Original-Revision: a50cea46aacf91f09c4243350de779ba136d7a43
GitOrigin-RevId: b24f8466e0512e3aea1375befdb374b1811336d2
Change-Id: Id38e2b011866047b7799c3420a1b7435659cb658