[roll] Roll fuchsia [a11y] InjectPointerEventAction: don't crash on null node
Early return instead of crashing when GetSemanticNode returns nullptr.

This is in response to an issue we were seeing in production:
fxbug.dev/110562
The issue likely still exists, but now we'll merely fail to process the
pass-through gesture (instead of crashing).

We've added an error log, so we can track this down in the future.

Original-Bug: 110562
Test: a11y-lib-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/733659
Original-Revision: fae3babef5f3ff7bf111d55c3d05afee28671e1a
GitOrigin-RevId: 2ea756898800c238f00d9ee1204453591be7abcf
Change-Id: I360253a50ecfc740fa1c210c3939f5f62e28b28e