[roll] Roll fuchsia [magma] Allow including headers as "magma/magma_*"
All magma headers are in the SDK under a "magma" directory, to help
reduce the risk of collisions with other libraries. We should make it
possible to include them using a the magma directory in-tree, to make it
easier to share code.

Original-Bug: 111274
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734507
Original-Revision: 16350fc2f1a8471f2baf5ce53f6a7d5af755148d
GitOrigin-RevId: 610d5efdffb8f52f483ca0a80ab4a2c7346c6f57
Change-Id: I20e68d7219c6eb58af03b6eed687986dc21767e1