[roll] Roll fuchsia [fidl][cpp] Flexible enums default initialize to Unknown
This is a prospective API change to be discussed during the
API review of the new C++ bindings.

It's more difficult to disallow default construction
of flexible enums because they are used in wire types which overall
is reliant on default construction.

Between zero or Unknown, Unknown is a slightly better choice because
it wouldn't accidentally send whatever element that happen to be
defined to zero.

RELNOTES: When default constructing flexible enums, note that they will
hold an unknown value instead of zero. Code that previously send default
constructed enums down the wire may now send a unknown value instead of
the member whose ABI is zero. To avoid sending unknowns, make sure
flexible enums are all properly initialized to a desired value instead
of relying on C++ default initialization.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726710
Original-Revision: 6fd01ab2d224084370d50036cc381db43e55c895
GitOrigin-RevId: 988ca7ad07ab23482a0edb2c0b2078faebb12a71
Change-Id: Ic2675b9b70ff2f4d457dec82597301a267764bd0