[roll] Roll fuchsia [component_manager] align resolution Context with pkg ResolutionContext
When the PackageResolver API was updated to support
ResolutionWithContext, the FIDL type declaration for the
ResolutionContext used a struct, with field `bytes`. This was an
improvement over the Component resolution::Context type that used an
alias for a byte array, because language bindings for the alias type
implemented the type as an array of bytes. The struct type in
PackageResolver is more type safe and more readable since the language
bindings always refer to the type as `ResolutionContext`.

This CL aligns the Component FIDL to use a similar struct declaration,
and updates API usage (such as in component_manager) accordingly.

Original-Bug: 100050

Original-Fixed: 104038

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735405
Original-Revision: b047371725fdb22d3f63ed236d0267c658630c45
GitOrigin-RevId: d1eab40f5665fddf0f5a4ec18ca5a9b662d1eb0c
Change-Id: I8a96ee85d19dd2034c65c2e0726e14baa2376173