[roll] Roll fuchsia [power][test] Ignore FIDL result in mock
Changes MockActivityService to ignore the result of calling
`on_state_changed` on the activity listener channel. Previously, an
error would cause a crash, which is not the correct behavior because the
service should be resilient to calling a method on a closed channel.

Original-Bug: 104266
Test: fx test //src/power/power-manager/integration
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735765
Original-Revision: 8a6dc8beee6f735333f8711a9d9c6be8da2b1e6f
GitOrigin-RevId: fbb724ada1906bdbd905a9bb854bcea74a2dfde7
Change-Id: Ia06f9777dfb231ca53e4042e6ad5b6b496f2671a