[roll] Roll fuchsia [sdmmc][fidl] Add method to re-enable the in-band interrupt
Add a method to ack the SDIO in-band interrupt to allow it to be
delivered again. SDMMC drivers will be changed to disable the interrupt
immediately after it is received, and only re-enable it after this
method has been called. The in-band interrupt is asserted by the card
until all interrupt sources have been cleared by function drivers, so
disabling it at the controller is needed to prevent a flood of
interrupts.

Original-Bug: 110355
Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729783
Original-Revision: afa5093beaadcf376b4a4847b7104498393e097e
GitOrigin-RevId: fca62f3a9805f90f018f4033e527f7a2b5a7de86
Change-Id: I124dad9866dca2468ac49b4896f4610f19f4fa9b