[roll] Roll fuchsia [sdio][fidl] Add method to re-enable the in-band interrupt
Add a method to ack the SDIO in-band interrupt to allow it to be
delivered again. The SDIO core driver will use this method to keep
interrupts disabled until function drivers have had a chance to handle
them. Explicitly ack'ing the interrupt is necessary to keep the core
driver and function drivers in sync and to prevent interrupts from being
missed.

Original-Bug: 110355
Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/731403
Original-Revision: 6c81e958226892252c0fa3475496612480eb931e
GitOrigin-RevId: 8a94ddaaafad5bb606b8c4a2a1918c9fc59f9b90
Change-Id: I0f125d976fee70a110ac7748474095ae6988f258