[roll] Roll fuchsia [sdio][core] Keep in-band interrupt disabled until a client handles it
An upcoming change will make SDMMC drivers disable the in-band interrupt
and leave it up to the SDIO core driver to re-enable it. Clearing the
interrupt source on the card is function-specific, so the best the core
driver can do is to re-enable the interrupt after a function driver
notifies it that the last one has been handled.

Original-Bug: 110355
Test: fx test sdmmc-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729903
Original-Revision: 6d6773de20c24a6dbeee54c96a5d8301412a07a7
GitOrigin-RevId: 7895c4274a4b0ff1870cc52cedc3aada3b7a2e4c
Change-Id: Iac606c4190f3e748d0b76694088bb709c26dc64a