[roll] Roll fuchsia [identity] Update all component log tags.
With the old sytle of diagnostics init each log statement included the
component in addition to the domain but recently we've lost that tag
which makes some of the logs hard to follow.

This CL adds the component name as a tag on all logs and also moves from
the ~2017 "auth" nomenclature to the "identity" we use in interfaces
and the source tree.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735406
Original-Revision: b7c0dd5ad1e76a68fe1ed59e4f1e22edcdf9c94e
GitOrigin-RevId: 25d95de0118ef9763045a8ccdf08a14c5ca42f93
Change-Id: I87181a89bbc0252051187051bf3d5cb464715ada