[roll] Roll fuchsia [input] Starts reporting AltGr as a modifier
The AltGr key is special because it is not defined in the USB HID.
So the only way an AltGr modifier can be reported is by examining
the KeyMeaning, which is not how our current modifier tracker
works.

For this reason, another modifier handler is required, which
fixes up the modifiers reported with the modifier bits
derived from key meanings.

At the moment, only AltGr is handled, since this was
immediately needed for Chromium.

Original-Bug: 110071, 109987
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728742
Original-Revision: 3b22ab31b73dc2bba74b93113ed4fe53a09cbdf1
GitOrigin-RevId: fcb8e936a11be4bdd74fe7697e619d723366e000
Change-Id: I54d3f933c79430e5fc6cc406043e0383bdfe3d3d