[roll] Roll fuchsia [subprocess] Fix race in test.
One of the tests attempts to kill a subprocess that doesn't finish on
its own even after receiving a SIGTERM. It does this by having the
subprocess execute a `sleep` during which it wouldn't respond to the
SIGTERM so that we can execute the path that waits for a grace period to
let the subprocess clean up and then sends a SIGKILL when the subprocess
fails to clean up in time. In the very rare occasion, this test would
flake because we'd cancel the context and send the SIGTERM before the
subprocess script runs `sleep` and it would just exit and clean up
normally. This change pgreps for the sleep process to make sure it was
run before we send the SIGTERM.

Test: Ran `fx go test -failfast -count=10000 ./tools/lib/subprocess`
twice.

Original-Bug: 111211
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734491
Original-Revision: 5603b4fda9a9ae4badacf2ca18c65a0f9130ac57
GitOrigin-RevId: c995f83e8c189f8b84711160df3ee37f2e6ac862
Change-Id: I4bf341df7a55e0d5fd2cb3f43e8a4ed0b1bde63a