[roll] Roll fuchsia [pb] Introduce pbtool which spits out a build archive
In order to facilitate the infra transition from build archives to
product bundles, we are introducing this tool which can take a product
bundle and spit out a build archive. This tool is only assumed to be
used by infra, therefore we do not want to add this to ffx. We are
writing this in Rust in order to share a lot of existing code with ffx.

Original-Bug: 110791
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/733942
Original-Revision: 4809040d89f13ef7be5d8446decf0a0f7a583ba4
GitOrigin-RevId: 7b48b209615da94b30a868977f996487c334ba07
Change-Id: Ib8f89cb7a5781c6b2fb95d32a90968c699968f29