[roll] Roll fuchsia [build] Fix expect_includes() for fuchsia_component_manifest().
When using fuchsia_component_manifest() before this change,
cmc_check_includes() never sees the binary with the appropriate metadata
as a dependency since the point of splitting up fuchsia_component() was
to allow the manifest to be compiled by a target which the binary can
dep on rather than being dep'd on. This means that when specifying
separate manifest & component targets the includes check is always a
no-op.

This moves the include checking into fuchsia_component() rather than
the manifest processor target so that the appropriate binaries will
always be present in deps.

In order for fuchsia_component() to invoke cmc_check_includes() it
must know the source path to the component manifest, but when using
fuchsia_component_manifest() the source path is not directly visible
to fuchsia_component(). This change adds the manifest path as metadata
which is collected by fuchsia_component() and adds a script to read the
metadata and run cmc with the actual manifest path.

Note that this change also adds some shards to manifests which
previously slipped through the checks. A future CL will clean up the
extra inspect shards once structured config client libraries have the
correct inspect dependency.

Original-Bug: 109273
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734509
Original-Revision: 3383a7b601c5ca28eaff690b9694738c229ed6b4
GitOrigin-RevId: 76c748b128b0d8a49d81750867ac858688e4af47
Change-Id: I1eaa768181aa457c3e9982f4a339d9b95971c87f