[roll] Roll fuchsia [input] Add a TODO to remove @transitional
This will allow us to soft-migrate its use in Flutter in a saner way.

I originally also wanted to add @available(added=10), but the lack
of annotation when this was initially merged means that the API
was already implicitly added to API level 9. So, can't modify that.

Canonical docs are here: https://fuchsia.dev/fuchsia-src/reference/fidl/language/versioning?hl=en, but I'm not sure they will show you this particular gadget. It's a consequence of how versioning works.

When the FIDL compiler runs, it gets passed an API level N. Any declarations that do not apply to API level N are not compiled.

The effect of these two declarations is that at API level 9 and smaller, the above definition applies. At 10 or larger, the below definition applies.
Net result is, up to 9 the method has @transitional, at 10+ the method does not
have @transitional.

FIDL folks are working on syntactic sugar that would allow us to collapse these two definitions into one, but it is not ready yet.

Original-Bug: 111291
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735752
Original-Revision: b862c5f0f1ed53a68926db4de6c6fc133050946a
GitOrigin-RevId: c66642a2c8f08804fbb33eab3fc56e3919b68d0a
Change-Id: If18feba27e33dfa80af3a976fbb1a8e07e499d5c