[roll] Roll fuchsia [debugger] Replace the "actions" system.
The actions system runs a sequence of commands. It is used for the
started attach/connect switches, and also a custom file that users can
provide to run some startup commands.

The way the actions work is that it's difficult to convert to the new
CommandContext system, especially the tests. It's also relatively
complex for what it does.

This patch replaces it with a simple function that chains the callbacks.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/730005
Original-Revision: 9dbb87e4a4a5fc045865fbaa00de4fe2c0396cc3
GitOrigin-RevId: 68b3be0332f0a75d1a56df81a765ff209355b6ef
Change-Id: Id7f465247cc0b6dd7603e28b103afda6cb871bd8