[roll] Roll fuchsia [netstack3] `shutdown` and `close` for TCP core sockets
Similar to posix `shutdown` and `close`: `shutdown` queues a FIN for a
connection; `close` first does `shutdown`, but also removes the socket
from the socket map if the connection becomes `Closed`.

Note that SHUT_RD is not implemented in Core, because it is not visible
from the TCP state machine itself, and usually have divergent behaviors
on different platforms, we will handle it in Bindings.

Test: fx test netstack3-core-test-package

Original-Fixed: 110573
Original-Bug: 98970, 103979
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735285
Original-Revision: 130bd518856ee4bbed2a0282c018859893dd9220
GitOrigin-RevId: cff671e1c5736dcdca37fd63e7b441c0a6b29e36
Change-Id: I8d125e30b9815ab68b90aa045aee6861ff946ac6