[roll] Roll fuchsia [wlan][policy] Don't expose SME for scanning.
When scanning or during network selection, the IfaceManager
hands out a raw client SME for other clients to use to
perform scans.  While the SME is intended to be used only
for scanning, it could be used to change the client
connectivity state as well.

This change wraps the SME in a struct that only allows for
the scan command to be passed through to the SME.  It also
allows us to store some additional metadata about the
client interface associated with the SME (namely the ID)
to be used in the future for reporting events from which we
would like to recover.

Original-Bug: 60089
Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734668
Original-Revision: 3660ab8988bb98a58a169e421fa1a71bb00cb2cd
GitOrigin-RevId: 201cc0579ef59ddbb85df24e25d321ceca4c3611
Change-Id: I6a293d7578ec79a17ba5f1343870313d5fd0edee