[roll] Roll fuchsia [storage-manager] AccountManager tests should generate StorageManager objects by factory.
A future change will expect AccountManager to accept a lambda which can
generate a struct which implements StorageManager, rather than a single
struct which implements StorageManager. This CL goes through all of the
testing callsites to perform that refactor eagerly, so that review of
that change is more self-contained.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735842
Original-Revision: aac060beecf800562971dbde20edbf98ddd0ed8b
GitOrigin-RevId: 930172793b76f78e8c10f167507729097dae29b9
Change-Id: Ic0bebc026bf90d9524e1f20ed1c526e5becb07f5