[roll] Roll fuchsia [spi] Add SPI Banjo device
Similar to fxr/713442, the old arrangement where the SPI device is both
bindable and adds a FIDL child to itself doesn't work in DFv2. So
instead we make a separate Banjo device that is the child of the SPI
device and the sibling of the FIDL device.

Test: booted Chromebook on DFv2 and verified that SPI devices now work
Original-Fixed: 108557
Original-Bug: 103068
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/736123
Original-Revision: ecce44ee71e88a98f3507a018aa79a559e8307ea
GitOrigin-RevId: 809ded71171d1ca65709bdf84bc607314b98095b
Change-Id: I455a98e269baa4f853566d1f981f791a95cffb40