[roll] Roll fuchsia [nxpfmac] Disconnect in connection destructor
The ClientConnection destructor would cancel any ongoing connection
attempts and make sure that any ongoing disconnects would complete. But
it would not actually perform a disconnect if it was connected.

Make sure we revert to a disconnected state if a client connection
object is destroyed. By waiting for the disconnect in progress state we
ensure that the disconnect either completed or wasn't necessary.

Test: Unit tests, acts start/stop client tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735866
Original-Revision: cbcae97b269d19a79676c84d5576edc3ae3a87c5
GitOrigin-RevId: c787a0c1ca1e066665aa9c86748ea3e29148ab9b
Change-Id: If27bb2aa36a874da06da6c32502b3f5cbee6eea7