[roll] Roll fuchsia [fidl][scripts] Extend scaffolding create command
The previous change added the scaffolding script, and endowed it with
one command: "create", which (as of that chagne) only works to the end
of creating a *new* canonical example series' baseline case. This change
extends the create command to also allow for the creation of variants on
an already existing series. The `scaffold.py create` invocation must
take either a `--new` flag (indicating a new series), or a
`--from=SERIES` flag, indicating a variant on an existing series.

Test: manual, used to generate fxr/737545
Original-Bug: 111614
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/737209
Original-Revision: 04797ebc8131b3592ab7cee5e18a378d815ad27d
GitOrigin-RevId: 18dd4245708f6b1361d8564b4012c7c111a05bb3
Change-Id: I0367902d1b8e2f867ee0ba5a706486e391038134