[roll] Roll fuchsia [fidl][scripts] Add `scaffold document` command
This command takes an already created canonical example variant and adds
documentation to it. That doesn't mean it writes the actual
documentation (that would be quite a feat...), but it does create all of
the documentation templates (like devsite widgets, index entries, etc),
and leaves `TODO(DONOTSUBMIT)` comments for the user. The output of the
command then tells the user where all of these comments are located, so
that they may replace them with the appropriate text.

Test: manual, used to generate fxr/737546 (PS1)
Original-Bug: 111614
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/737210
Original-Revision: 21311c24d0353e67a6dce04fc2509037e45fff21
GitOrigin-RevId: 06b1ca52ad5ee0ce3d54067a5fd42af762eb346d
Change-Id: I63deacbfc9d3d3ce3167d1812316fa38517f7744