[roll] Roll fuchsia [fidl][scripts] Add `scaffold implement` for Rust
This command takes an already created canonical example variant and adds
an implementation to it. Additionally, if this is the first
implementation for the canonical example variant in question, it adds a
test implementation as well. This change only implements this
functionality for the /rust and /test directories - the follow up
implementation will do the same for all of the C++ directories
(/cpp_wire, /cpp_natural, and /hlcpp) as well.

Test: manual, used to generate fxr/737547
Original-Bug: 111614
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/737543
Original-Revision: 9d36bf13edae99aaaa591e6e0d289fc268ace68d
GitOrigin-RevId: 422260ae648771f7e77d2c9bbaedc1aa4abbb6d1
Change-Id: Ib909bccbe0ce3234fca507300de73e61c6386de8