[roll] Roll fuchsia [pbm] Fix errors on pb remove with no bundles
The attached bug reports on an error condition when removing
product-bundles if none are in the cache. This is just an uncaught error
mode when the list is empty, which is unhandled by the tool.

This CL explicitly handles this error condition and outputs it as
information rather than an error.

Original-Bug: 111356
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/736162
Original-Revision: e0713057d0a7c24d44e724076f3df30e4ac422b4
GitOrigin-RevId: 906926ea9a2774ca9771f5c8f64ef5ed56bd368f
Change-Id: Ib892aed3d67e1451d153e50ff88cca12c7971f95