[roll] Roll fuchsia [ui] Change fake display dimensions.
Currently the fake display dimensions are set as 1024X600. The value of
the height is an issue since it is not divisible by 32 and creates
precision errors in ui integration tests. This CL changes the
dimesnsions to 1280X800 which makes testing easier.

Original-Bug: 111483

Test: CQ

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/736128
Original-Revision: 219119ab034a147d99558937dee17057cc2271c4
GitOrigin-RevId: f5910eae39a4462ad22c7be9017efade572c33d8
Change-Id: Ia67e2dea8b1af56a3810bd10c7138393db937064